Mohr circle parameters

calculates parameters of the Mohr circle

Usage

Mohr_calc(
  sigma_x = NA,
  sigma_z = NA,
  tau_xz = NA,
  sigma1 = NA,
  sigma3 = NA,
  theta = seq(from = 0, to = 180, by = 1)
)

Arguments

sigma_x

numeric. Magnitude of normal stress acting in the horizontal direction

sigma_z

numeric. Magnitude of normal stress acting in the vertical direction

tau_xz

numeric. Magnitude of shear stress acting on the same plane as "sigma_x"

sigma1

numeric. Magnitude of major principal stress

sigma3

numeric. Magnitude of minor principal stress

theta

numeric. Angles (degrees); defaults to 0-180 in increments of 1

Value

A five-element list containing

"mean"

center of Mohr circle, i.e. the mean stress

"deviatoric"

radius of Mohr circle, i.e. the deviatoric stress

"sigma"

vector of normal stresses for Mohr circle

"tau"

vector of shear stresses for Mohr circle

"theta"

vector of angles (degrees)

Details

If theta is entered in conjunction with "sigma_x", "sigma_z", and "tau_xz", it is interpreted as the angle of inclination above the horizontal. If "theta" is entered in conjunction with the principal stresses, then it is interpreted as the angle of inclination above the major principal plane.

Note

One of the following two sets of data must be entered

"sigma_x", "sigma_z", "tau_xz"

"sigma1", "sigma3"

See also

Author

Kyle Elmy and Jim Kaklamanos

Examples

Mohr_calc(sigma_x = 80, sigma_z = 120, tau_xz = 20)
#> $mean
#> [1] 100
#> 
#> $deviatoric
#> [1] 28.28427
#> 
#> $normal
#>   [1] 120.00000 120.34600 120.68581 121.01931 121.34641 121.66701 121.98101
#>   [8] 122.28831 122.58882 122.88246 123.16912 123.44872 123.72119 123.98642
#>  [15] 124.24435 124.49490 124.73798 124.97353 125.20147 125.42173 125.63426
#>  [22] 125.83897 126.03581 126.22472 126.40564 126.57852 126.74330 126.89994
#>  [29] 127.04838 127.18859 127.32051 127.44411 127.55935 127.66619 127.76461
#>  [36] 127.85457 127.93604 128.00901 128.07344 128.12933 128.17664 128.21537
#>  [43] 128.24551 128.26704 128.27996 128.28427 128.27996 128.26704 128.24551
#>  [50] 128.21537 128.17664 128.12933 128.07344 128.00901 127.93604 127.85457
#>  [57] 127.76461 127.66619 127.55935 127.44411 127.32051 127.18859 127.04838
#>  [64] 126.89994 126.74330 126.57852 126.40564 126.22472 126.03581 125.83897
#>  [71] 125.63426 125.42173 125.20147 124.97353 124.73798 124.49490 124.24435
#>  [78] 123.98642 123.72119 123.44872 123.16912 122.88246 122.58882 122.28831
#>  [85] 121.98101 121.66701 121.34641 121.01931 120.68581 120.34600 120.00000
#>  [92] 119.64791 119.28983 118.92587 118.55615 118.18078 117.79987 117.41354
#>  [99] 117.02190 116.62508 116.22319 115.81636 115.40472 114.98838 114.56748
#> [106] 114.14214 113.71249 113.27866 112.84079 112.39901 111.95345 111.50425
#> [113] 111.05155 110.59547 110.13618 109.67379 109.20846 108.74032 108.26952
#> [120] 107.79620 107.32051 106.84258 106.36258 105.88063 105.39689 104.91151
#> [127] 104.42463 103.93641 103.44699 102.95651 102.46514 101.97301 101.48028
#> [134] 100.98711 100.49363 100.00000  99.50637  99.01289  98.51972  98.02699
#> [141]  97.53486  97.04349  96.55301  96.06359  95.57537  95.08849  94.60311
#> [148]  94.11937  93.63742  93.15742  92.67949  92.20380  91.73048  91.25968
#> [155]  90.79154  90.32621  89.86382  89.40453  88.94845  88.49575  88.04655
#> [162]  87.60099  87.15921  86.72134  86.28751  85.85786  85.43252  85.01162
#> [169]  84.59528  84.18364  83.77681  83.37492  82.97810  82.58646  82.20013
#> [176]  81.81922  81.44385  81.07413  80.71017  80.35209  80.00000
#> 
#> $shear
#>   [1] -2.000000e+01 -1.964791e+01 -1.928983e+01 -1.892587e+01 -1.855615e+01
#>   [6] -1.818078e+01 -1.779987e+01 -1.741354e+01 -1.702190e+01 -1.662508e+01
#>  [11] -1.622319e+01 -1.581636e+01 -1.540472e+01 -1.498838e+01 -1.456748e+01
#>  [16] -1.414214e+01 -1.371249e+01 -1.327866e+01 -1.284079e+01 -1.239901e+01
#>  [21] -1.195345e+01 -1.150425e+01 -1.105155e+01 -1.059547e+01 -1.013618e+01
#>  [26] -9.673791e+00 -9.208458e+00 -8.740320e+00 -8.269521e+00 -7.796202e+00
#>  [31] -7.320508e+00 -6.842585e+00 -6.362577e+00 -5.880631e+00 -5.396893e+00
#>  [36] -4.911512e+00 -4.424635e+00 -3.936410e+00 -3.446986e+00 -2.956511e+00
#>  [41] -2.465137e+00 -1.973011e+00 -1.480284e+00 -9.871068e-01 -4.936286e-01
#>  [46] -1.776357e-15  4.936286e-01  9.871068e-01  1.480284e+00  1.973011e+00
#>  [51]  2.465137e+00  2.956511e+00  3.446986e+00  3.936410e+00  4.424635e+00
#>  [56]  4.911512e+00  5.396893e+00  5.880631e+00  6.362577e+00  6.842585e+00
#>  [61]  7.320508e+00  7.796202e+00  8.269521e+00  8.740320e+00  9.208458e+00
#>  [66]  9.673791e+00  1.013618e+01  1.059547e+01  1.105155e+01  1.150425e+01
#>  [71]  1.195345e+01  1.239901e+01  1.284079e+01  1.327866e+01  1.371249e+01
#>  [76]  1.414214e+01  1.456748e+01  1.498838e+01  1.540472e+01  1.581636e+01
#>  [81]  1.622319e+01  1.662508e+01  1.702190e+01  1.741354e+01  1.779987e+01
#>  [86]  1.818078e+01  1.855615e+01  1.892587e+01  1.928983e+01  1.964791e+01
#>  [91]  2.000000e+01  2.034600e+01  2.068581e+01  2.101931e+01  2.134641e+01
#>  [96]  2.166701e+01  2.198101e+01  2.228831e+01  2.258882e+01  2.288246e+01
#> [101]  2.316912e+01  2.344872e+01  2.372119e+01  2.398642e+01  2.424435e+01
#> [106]  2.449490e+01  2.473798e+01  2.497353e+01  2.520147e+01  2.542173e+01
#> [111]  2.563426e+01  2.583897e+01  2.603581e+01  2.622472e+01  2.640564e+01
#> [116]  2.657852e+01  2.674330e+01  2.689994e+01  2.704838e+01  2.718859e+01
#> [121]  2.732051e+01  2.744411e+01  2.755935e+01  2.766619e+01  2.776461e+01
#> [126]  2.785457e+01  2.793604e+01  2.800901e+01  2.807344e+01  2.812933e+01
#> [131]  2.817664e+01  2.821537e+01  2.824551e+01  2.826704e+01  2.827996e+01
#> [136]  2.828427e+01  2.827996e+01  2.826704e+01  2.824551e+01  2.821537e+01
#> [141]  2.817664e+01  2.812933e+01  2.807344e+01  2.800901e+01  2.793604e+01
#> [146]  2.785457e+01  2.776461e+01  2.766619e+01  2.755935e+01  2.744411e+01
#> [151]  2.732051e+01  2.718859e+01  2.704838e+01  2.689994e+01  2.674330e+01
#> [156]  2.657852e+01  2.640564e+01  2.622472e+01  2.603581e+01  2.583897e+01
#> [161]  2.563426e+01  2.542173e+01  2.520147e+01  2.497353e+01  2.473798e+01
#> [166]  2.449490e+01  2.424435e+01  2.398642e+01  2.372119e+01  2.344872e+01
#> [171]  2.316912e+01  2.288246e+01  2.258882e+01  2.228831e+01  2.198101e+01
#> [176]  2.166701e+01  2.134641e+01  2.101931e+01  2.068581e+01  2.034600e+01
#> [181]  2.000000e+01
#> 
#> $theta
#>   [1]   0   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17
#>  [19]  18  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35
#>  [37]  36  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53
#>  [55]  54  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71
#>  [73]  72  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89
#>  [91]  90  91  92  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107
#> [109] 108 109 110 111 112 113 114 115 116 117 118 119 120 121 122 123 124 125
#> [127] 126 127 128 129 130 131 132 133 134 135 136 137 138 139 140 141 142 143
#> [145] 144 145 146 147 148 149 150 151 152 153 154 155 156 157 158 159 160 161
#> [163] 162 163 164 165 166 167 168 169 170 171 172 173 174 175 176 177 178 179
#> [181] 180
#>