Principal stresses from 2D stress components

Determines the principal stresses and their orientations from the stress components $$sigma_x$$, $$\sigma_y$$, $$\tau_{xy}$$.

Usage

PR_stress(sigmaX, sigmaY, tauXY)

Arguments

sigmaX

numeric. Magnitude of normal stress acting in the horizontal direction

sigmaY

numeric. Magnitude of normal stress acting on plane facing in Y direction ($$\sigma_y$$).

tauXY

numeric. Magnitude of shear stress acting on planes facing X and Y ($$\tau_{xy}$$).

Value

angle in degrees

References

Richard J. Lisle (1999)