Confidence ellipse

Bootstrapped projected mean with percentile confidence region and hypothesis tests

Usage

confidence_ellipse(x, n = 10000L, alpha = 0.05, res = 512L, isotropic = FALSE)

Source

geologyGeometry (J.R. Davis)

Arguments

x

object of class "Vec3", "Line", or "Plane"

n

integer (10000 by default).

alpha

numeric (0.05 by default).

res

integer. The resolution with which to sample the boundary curve of the (1 - alpha) * 100% percentile region.

isotropic

logical. If TRUE, forces the inverse covariance to the identity matrix and hence the region to be circular.

Value

list.

center

Projected mean of x

cov

Inverse covariance matrix in the tangent space, which is the identity if isotropic is TRUE

rotation

Rotation matrix used

quantiles

Quantiles of Mahalanobis norm

pvalue.ray

For rays: p-value for each ray in x, i.e. the fraction of x that are farther from center than the given ray.

pvalue.line

For lines: p-value for each line in x, i.e. the fraction of x that are farther from center than the given line

pvalue.line.FUN,pvalue.ray.FUN

The function to calculate the p-value for a given vector

angles

Angles of the semi-axis of the confidence ellipse (in radians if x is an "Vec3" object, in degrees if otherwise.)

ellipse

Confidence ellipse given as "Vec3" object with res vectors

Note

The inference is based on percentiles of Mahalanobis distance in the tangent space at the mean of the bootstrapped means. The user should check that the bootstrapped means form a tight ellipsoidal cluster, before taking such a region seriously.

References

Davis, J. R., & Titus, S. J. (2017). Modern methods of analysis for three-dimensional orientational data. Journal of Structural Geology, 96, 65–89. https://doi.org/10.1016/j.jsg.2017.01.002

Examples

set.seed(20250411)
ce <- confidence_ellipse(example_lines, n = 1000, res = 10)
print(ce)
#> $center
#> Line object (n = 1):
#>  azimuth   plunge 
#> 69.19989 14.80455 
#> 
#> $cov
#>           [,1]      [,2]
#> [1,] 1265.7702 -268.7996
#> [2,] -268.7996 1773.4826
#> 
#> $rotation
#>              [,1]       [,2]       [,3]
#> center  0.3433202 0.90379175  0.2555225
#> perp   -0.6987677 0.42757927 -0.5734977
#>        -0.6275786 0.01834249  0.7783371
#> attr(,"class")
#> [1] "Rotation" "matrix"   "array"   
#> 
#> $quantiles
#>            q000      q025     q050     q075     q095     q099     q100
#> [1,] 0.01479835 0.7477178 1.148652 1.675819 2.489669 3.016634 4.225523
#> 
#> $pvalue.ray.FUN
#> function (u) 
#> {
#>     if (is.spherical(u)) 
#>         u <- unclass(Vec3(u))
#>     v <- rayTangentVectorFromPoint(u, rot)
#>     1 - empiricalCDF(sqrt(v %*% covarInv %*% v))
#> }
#> <bytecode: 0x561535bff608>
#> <environment: 0x561535bfdb80>
#> 
#> $pvalue.ray
#>    [1] 0.162 0.401 0.831 0.570 0.665 0.438 0.621 0.989 0.791 0.482 0.344 0.010
#>   [13] 0.058 0.101 0.446 0.427 0.884 0.171 0.644 0.060 0.190 0.321 0.947 0.152
#>   [25] 0.874 0.865 0.109 0.118 0.659 0.851 0.542 0.233 0.018 0.601 0.365 0.735
#>   [37] 0.766 0.129 0.454 0.398 0.368 0.683 0.607 0.950 0.155 0.626 0.272 0.337
#>   [49] 0.921 0.514 0.267 0.879 0.185 0.177 0.767 0.990 0.527 0.736 0.711 0.501
#>   [61] 0.551 0.513 0.705 0.964 0.372 0.289 0.237 0.071 0.404 0.074 0.132 0.658
#>   [73] 0.439 0.055 0.983 0.226 0.775 0.688 0.136 0.493 0.955 0.583 0.779 0.473
#>   [85] 0.214 0.842 0.232 0.843 0.145 0.378 0.746 0.941 0.785 0.351 0.783 0.862
#>   [97] 0.085 0.224 0.193 0.515 0.958 0.369 0.383 0.832 0.860 0.877 0.671 0.432
#>  [109] 0.306 0.778 0.708 0.603 0.636 0.772 0.824 0.952 0.739 0.646 0.488 0.093
#>  [121] 0.771 0.412 0.940 0.213 0.837 0.292 0.946 0.329 0.502 0.212 0.980 0.141
#>  [133] 0.235 0.078 0.814 0.263 0.800 0.134 0.324 0.889 0.684 0.539 0.925 0.420
#>  [145] 0.755 0.627 0.070 0.481 0.367 0.859 0.526 0.619 0.654 0.793 0.022 0.215
#>  [157] 0.593 0.901 0.002 0.080 0.787 0.032 0.031 0.945 0.460 0.552 0.682 0.571
#>  [169] 0.555 0.942 0.809 0.528 0.937 0.608 0.727 0.096 0.511 0.961 0.107 0.414
#>  [181] 0.305 0.547 0.323 0.935 0.686 0.221 0.931 0.722 0.186 0.700 0.105 0.153
#>  [193] 0.597 0.380 0.704 0.647 0.068 0.690 0.857 0.381 0.580 0.024 0.459 0.220
#>  [205] 0.794 0.948 0.668 0.194 0.490 0.680 0.795 0.312 0.805 0.142 0.852 0.765
#>  [217] 0.126 0.023 0.826 0.286 0.624 0.130 0.834 0.082 0.339 0.326 0.495 0.956
#>  [229] 0.576 0.335 0.057 0.652 0.334 0.676 0.716 0.276 0.531 0.199 0.089 0.349
#>  [241] 0.768 0.776 0.176 0.362 0.125 0.163 0.150 0.167 0.415 0.418 0.348 0.585
#>  [253] 0.284 0.434 0.356 0.385 0.744 0.435 0.474 0.503 0.914 0.005 0.781 0.029
#>  [265] 0.191 0.492 0.886 0.618 0.317 0.804 0.602 0.067 0.346 0.411 0.664 0.715
#>  [277] 0.737 0.475 0.866 0.604 0.748 0.160 0.366 0.149 0.308 0.828 0.541 0.059
#>  [289] 0.449 0.638 0.140 0.995 0.757 0.954 0.480 0.006 0.639 0.448 0.281 0.259
#>  [301] 0.403 0.616 0.165 0.230 0.077 0.393 0.003 0.559 0.613 0.634 0.588 0.379
#>  [313] 0.433 0.953 0.218 0.713 0.487 0.178 0.196 0.181 0.614 0.117 0.094 0.425
#>  [325] 0.903 0.734 0.529 0.016 0.764 0.965 0.370 0.525 0.421 0.115 0.536 0.885
#>  [337] 0.316 0.166 0.248 0.204 0.724 0.327 0.904 0.437 0.228 0.957 0.996 0.718
#>  [349] 0.173 0.825 0.044 0.615 0.081 0.717 0.236 0.573 0.817 0.251 0.894 0.761
#>  [361] 0.046 0.788 0.721 0.187 0.909 0.354 0.039 0.625 0.892 0.104 0.848 0.510
#>  [373] 0.978 0.699 0.919 0.358 0.698 0.100 0.750 0.428 0.537 0.898 0.975 0.069
#>  [385] 0.011 0.908 0.899 0.709 0.591 0.328 0.543 0.396 0.238 0.590 0.916 0.465
#>  [397] 0.279 0.461 0.269 0.436 0.315 0.628 0.304 0.314 0.938 0.747 0.623 0.972
#>  [409] 0.035 0.694 0.331 0.359 0.205 0.360 0.803 0.592 0.632 0.520 0.612 0.357
#>  [421] 0.034 0.810 0.470 0.936 0.882 0.164 0.929 0.198 0.182 0.774 0.458 0.217
#>  [433] 0.462 0.743 0.560 0.661 0.841 0.906 0.397 0.532 0.813 0.406 0.174 0.192
#>  [445] 0.984 0.920 0.299 0.066 0.361 0.728 0.250 0.657 0.062 0.268 0.498 0.666
#>  [457] 0.533 0.252 0.447 0.733 0.887 0.169 0.311 0.264 0.064 0.409 0.128 0.900
#>  [469] 0.451 0.753 0.210 0.796 0.287 0.867 0.322 0.855 0.225 0.040 0.245 0.981
#>  [481] 0.229 0.821 0.912 0.270 0.222 0.133 0.131 0.020 0.789 0.740 0.780 0.584
#>  [493] 0.333 0.239 0.076 0.387 0.924 0.838 0.522 0.274 0.509 0.594 0.719 0.600
#>  [505] 0.575 0.219 0.928 0.836 0.469 0.061 0.849 0.992 0.674 0.663 0.202 0.905
#>  [517] 0.569 0.696 0.792 0.745 0.758 0.336 0.564 0.389 0.243 0.786 0.120 0.726
#>  [529] 0.376 0.135 0.386 0.175 0.725 0.201 0.963 0.200 0.330 0.922 0.629 0.296
#>  [541] 0.122 0.518 0.959 0.206 0.041 0.244 0.463 0.103 0.188 0.987 0.998 0.402
#>  [553] 0.566 0.697 0.399 0.293 0.823 0.216 0.265 0.521 0.374 0.808 0.563 0.053
#>  [565] 0.318 0.507 0.485 0.209 0.648 0.504 0.442 0.662 0.283 0.681 0.907 0.723
#>  [577] 0.075 0.303 0.391 0.278 0.943 0.752 0.410 0.534 0.872 0.967 0.203 0.231
#>  [589] 0.464 0.524 0.782 0.030 0.962 0.911 0.806 0.424 0.013 0.530 0.609 0.025
#>  [601] 0.352 0.572 0.701 0.027 0.893 0.364 0.275 0.548 0.407 0.102 0.982 0.302
#>  [613] 0.620 0.516 0.499 0.240 0.693 0.091 0.083 0.241 0.065 0.582 0.147 0.720
#>  [625] 0.280 0.707 0.738 0.483 0.988 0.486 0.468 0.650 0.902 0.137 0.562 0.505
#>  [637] 0.635 0.072 0.567 0.557 0.189 0.431 0.976 0.986 0.417 0.784 0.373 0.999
#>  [649] 0.799 0.994 0.545 0.484 0.042 0.113 0.262 0.050 0.043 0.970 0.549 0.558
#>  [661] 0.496 0.815 0.384 0.054 0.161 0.926 0.883 0.291 0.589 0.660 0.847 0.606
#>  [673] 0.004 0.148 0.643 0.861 0.910 0.812 0.478 0.586 0.234 0.840 0.869 0.310
#>  [685] 0.110 0.816 0.341 0.494 0.112 0.556 0.692 0.088 0.732 0.021 0.049 0.026
#>  [697] 0.971 0.844 0.642 0.466 0.260 0.108 0.895 0.622 0.633 0.610 0.000 0.395
#>  [709] 0.144 0.345 0.710 0.340 0.208 0.598 0.008 0.157 0.871 0.760 0.319 0.687
#>  [721] 0.457 0.669 0.452 0.271 0.568 0.489 0.400 0.801 0.254 0.997 0.827 0.677
#>  [733] 0.811 0.819 0.645 0.540 0.114 0.993 0.672 0.290 0.807 0.123 0.881 0.332
#>  [745] 0.508 0.444 0.325 0.146 0.769 0.839 0.579 0.172 0.756 0.413 0.445 0.266
#>  [757] 0.875 0.544 0.298 0.001 0.119 0.301 0.870 0.116 0.467 0.179 0.651 0.691
#>  [769] 0.036 0.933 0.430 0.685 0.471 0.195 0.012 0.388 0.888 0.246 0.863 0.830
#>  [781] 0.864 0.295 0.183 0.553 0.731 0.655 0.673 0.675 0.282 0.342 0.455 0.932
#>  [793] 0.574 0.856 0.565 0.143 0.253 0.951 0.347 0.211 0.472 0.822 0.679 0.017
#>  [805] 0.850 0.073 0.695 0.702 0.009 0.377 0.798 0.138 0.223 0.587 0.158 0.052
#>  [817] 0.968 0.121 0.363 0.184 0.667 0.519 0.868 0.546 0.578 0.168 0.242 0.829
#>  [829] 0.802 0.450 0.390 0.985 0.151 0.063 0.441 0.535 0.561 0.977 0.641 0.277
#>  [841] 0.440 0.729 0.854 0.846 0.927 0.611 0.051 0.255 0.090 0.477 0.500 0.730
#>  [853] 0.678 0.759 0.258 0.124 0.712 0.285 0.419 0.878 0.084 0.300 0.749 0.033
#>  [865] 0.790 0.751 0.517 0.247 0.873 0.156 0.408 0.092 0.550 0.127 0.456 0.949
#>  [877] 0.880 0.512 0.858 0.375 0.353 0.453 0.918 0.382 0.637 0.087 0.915 0.991
#>  [889] 0.599 0.197 0.019 0.429 0.966 0.835 0.649 0.762 0.443 0.426 0.056 0.261
#>  [901] 0.670 0.405 0.538 0.714 0.876 0.079 0.689 0.577 0.960 0.479 0.111 0.741
#>  [913] 0.249 0.227 0.139 0.047 0.313 0.930 0.853 0.086 0.297 0.923 0.037 0.974
#>  [925] 0.095 0.098 0.845 0.703 0.416 0.770 0.595 0.392 0.497 0.288 0.097 0.754
#>  [937] 0.014 0.653 0.605 0.423 0.656 0.706 0.581 0.180 0.820 0.350 0.506 0.343
#>  [949] 0.106 0.007 0.338 0.640 0.028 0.045 0.048 0.309 0.099 0.038 0.934 0.890
#>  [961] 0.523 0.917 0.015 0.307 0.797 0.763 0.773 0.371 0.973 0.777 0.159 0.631
#>  [973] 0.833 0.944 0.257 0.476 0.554 0.913 0.596 0.294 0.939 0.256 0.630 0.891
#>  [985] 0.979 0.818 0.355 0.273 0.320 0.742 0.896 0.491 0.422 0.394 0.154 0.170
#>  [997] 0.617 0.969 0.897 0.207
#> 
#> $pvalue.line.FUN
#> function (l) 
#> {
#>     if (is.spherical(l)) 
#>         l <- unclass(Vec3(l))
#>     if (dot(l, inf$center) < 0) {
#>         inf$pvalue(-l)
#>     }
#>     else {
#>         inf$pvalue(l)
#>     }
#> }
#> <bytecode: 0x5615347e2c38>
#> <environment: 0x5615347e3ed0>
#> 
#> $pvalue.line
#>    [1] 0.162 0.401 0.831 0.570 0.665 0.438 0.621 0.989 0.791 0.482 0.344 0.010
#>   [13] 0.058 0.101 0.446 0.427 0.884 0.171 0.644 0.060 0.190 0.321 0.947 0.152
#>   [25] 0.874 0.865 0.109 0.118 0.659 0.851 0.542 0.233 0.018 0.601 0.365 0.735
#>   [37] 0.766 0.129 0.454 0.398 0.368 0.683 0.607 0.950 0.155 0.626 0.272 0.337
#>   [49] 0.921 0.514 0.267 0.879 0.185 0.177 0.767 0.990 0.527 0.736 0.711 0.501
#>   [61] 0.551 0.513 0.705 0.964 0.372 0.289 0.237 0.071 0.404 0.074 0.132 0.658
#>   [73] 0.439 0.055 0.983 0.226 0.775 0.688 0.136 0.493 0.955 0.583 0.779 0.473
#>   [85] 0.214 0.842 0.232 0.843 0.145 0.378 0.746 0.941 0.785 0.351 0.783 0.862
#>   [97] 0.085 0.224 0.193 0.515 0.958 0.369 0.383 0.832 0.860 0.877 0.671 0.432
#>  [109] 0.306 0.778 0.708 0.603 0.636 0.772 0.824 0.952 0.739 0.646 0.488 0.093
#>  [121] 0.771 0.412 0.940 0.213 0.837 0.292 0.946 0.329 0.502 0.212 0.980 0.141
#>  [133] 0.235 0.078 0.814 0.263 0.800 0.134 0.324 0.889 0.684 0.539 0.925 0.420
#>  [145] 0.755 0.627 0.070 0.481 0.367 0.859 0.526 0.619 0.654 0.793 0.022 0.215
#>  [157] 0.593 0.901 0.002 0.080 0.787 0.032 0.031 0.945 0.460 0.552 0.682 0.571
#>  [169] 0.555 0.942 0.809 0.528 0.937 0.608 0.727 0.096 0.511 0.961 0.107 0.414
#>  [181] 0.305 0.547 0.323 0.935 0.686 0.221 0.931 0.722 0.186 0.700 0.105 0.153
#>  [193] 0.597 0.380 0.704 0.647 0.068 0.690 0.857 0.381 0.580 0.024 0.459 0.220
#>  [205] 0.794 0.948 0.668 0.194 0.490 0.680 0.795 0.312 0.805 0.142 0.852 0.765
#>  [217] 0.126 0.023 0.826 0.286 0.624 0.130 0.834 0.082 0.339 0.326 0.495 0.956
#>  [229] 0.576 0.335 0.057 0.652 0.334 0.676 0.716 0.276 0.531 0.199 0.089 0.349
#>  [241] 0.768 0.776 0.176 0.362 0.125 0.163 0.150 0.167 0.415 0.418 0.348 0.585
#>  [253] 0.284 0.434 0.356 0.385 0.744 0.435 0.474 0.503 0.914 0.005 0.781 0.029
#>  [265] 0.191 0.492 0.886 0.618 0.317 0.804 0.602 0.067 0.346 0.411 0.664 0.715
#>  [277] 0.737 0.475 0.866 0.604 0.748 0.160 0.366 0.149 0.308 0.828 0.541 0.059
#>  [289] 0.449 0.638 0.140 0.995 0.757 0.954 0.480 0.006 0.639 0.448 0.281 0.259
#>  [301] 0.403 0.616 0.165 0.230 0.077 0.393 0.003 0.559 0.613 0.634 0.588 0.379
#>  [313] 0.433 0.953 0.218 0.713 0.487 0.178 0.196 0.181 0.614 0.117 0.094 0.425
#>  [325] 0.903 0.734 0.529 0.016 0.764 0.965 0.370 0.525 0.421 0.115 0.536 0.885
#>  [337] 0.316 0.166 0.248 0.204 0.724 0.327 0.904 0.437 0.228 0.957 0.996 0.718
#>  [349] 0.173 0.825 0.044 0.615 0.081 0.717 0.236 0.573 0.817 0.251 0.894 0.761
#>  [361] 0.046 0.788 0.721 0.187 0.909 0.354 0.039 0.625 0.892 0.104 0.848 0.510
#>  [373] 0.978 0.699 0.919 0.358 0.698 0.100 0.750 0.428 0.537 0.898 0.975 0.069
#>  [385] 0.011 0.908 0.899 0.709 0.591 0.328 0.543 0.396 0.238 0.590 0.916 0.465
#>  [397] 0.279 0.461 0.269 0.436 0.315 0.628 0.304 0.314 0.938 0.747 0.623 0.972
#>  [409] 0.035 0.694 0.331 0.359 0.205 0.360 0.803 0.592 0.632 0.520 0.612 0.357
#>  [421] 0.034 0.810 0.470 0.936 0.882 0.164 0.929 0.198 0.182 0.774 0.458 0.217
#>  [433] 0.462 0.743 0.560 0.661 0.841 0.906 0.397 0.532 0.813 0.406 0.174 0.192
#>  [445] 0.984 0.920 0.299 0.066 0.361 0.728 0.250 0.657 0.062 0.268 0.498 0.666
#>  [457] 0.533 0.252 0.447 0.733 0.887 0.169 0.311 0.264 0.064 0.409 0.128 0.900
#>  [469] 0.451 0.753 0.210 0.796 0.287 0.867 0.322 0.855 0.225 0.040 0.245 0.981
#>  [481] 0.229 0.821 0.912 0.270 0.222 0.133 0.131 0.020 0.789 0.740 0.780 0.584
#>  [493] 0.333 0.239 0.076 0.387 0.924 0.838 0.522 0.274 0.509 0.594 0.719 0.600
#>  [505] 0.575 0.219 0.928 0.836 0.469 0.061 0.849 0.992 0.674 0.663 0.202 0.905
#>  [517] 0.569 0.696 0.792 0.745 0.758 0.336 0.564 0.389 0.243 0.786 0.120 0.726
#>  [529] 0.376 0.135 0.386 0.175 0.725 0.201 0.963 0.200 0.330 0.922 0.629 0.296
#>  [541] 0.122 0.518 0.959 0.206 0.041 0.244 0.463 0.103 0.188 0.987 0.998 0.402
#>  [553] 0.566 0.697 0.399 0.293 0.823 0.216 0.265 0.521 0.374 0.808 0.563 0.053
#>  [565] 0.318 0.507 0.485 0.209 0.648 0.504 0.442 0.662 0.283 0.681 0.907 0.723
#>  [577] 0.075 0.303 0.391 0.278 0.943 0.752 0.410 0.534 0.872 0.967 0.203 0.231
#>  [589] 0.464 0.524 0.782 0.030 0.962 0.911 0.806 0.424 0.013 0.530 0.609 0.025
#>  [601] 0.352 0.572 0.701 0.027 0.893 0.364 0.275 0.548 0.407 0.102 0.982 0.302
#>  [613] 0.620 0.516 0.499 0.240 0.693 0.091 0.083 0.241 0.065 0.582 0.147 0.720
#>  [625] 0.280 0.707 0.738 0.483 0.988 0.486 0.468 0.650 0.902 0.137 0.562 0.505
#>  [637] 0.635 0.072 0.567 0.557 0.189 0.431 0.976 0.986 0.417 0.784 0.373 0.999
#>  [649] 0.799 0.994 0.545 0.484 0.042 0.113 0.262 0.050 0.043 0.970 0.549 0.558
#>  [661] 0.496 0.815 0.384 0.054 0.161 0.926 0.883 0.291 0.589 0.660 0.847 0.606
#>  [673] 0.004 0.148 0.643 0.861 0.910 0.812 0.478 0.586 0.234 0.840 0.869 0.310
#>  [685] 0.110 0.816 0.341 0.494 0.112 0.556 0.692 0.088 0.732 0.021 0.049 0.026
#>  [697] 0.971 0.844 0.642 0.466 0.260 0.108 0.895 0.622 0.633 0.610 0.000 0.395
#>  [709] 0.144 0.345 0.710 0.340 0.208 0.598 0.008 0.157 0.871 0.760 0.319 0.687
#>  [721] 0.457 0.669 0.452 0.271 0.568 0.489 0.400 0.801 0.254 0.997 0.827 0.677
#>  [733] 0.811 0.819 0.645 0.540 0.114 0.993 0.672 0.290 0.807 0.123 0.881 0.332
#>  [745] 0.508 0.444 0.325 0.146 0.769 0.839 0.579 0.172 0.756 0.413 0.445 0.266
#>  [757] 0.875 0.544 0.298 0.001 0.119 0.301 0.870 0.116 0.467 0.179 0.651 0.691
#>  [769] 0.036 0.933 0.430 0.685 0.471 0.195 0.012 0.388 0.888 0.246 0.863 0.830
#>  [781] 0.864 0.295 0.183 0.553 0.731 0.655 0.673 0.675 0.282 0.342 0.455 0.932
#>  [793] 0.574 0.856 0.565 0.143 0.253 0.951 0.347 0.211 0.472 0.822 0.679 0.017
#>  [805] 0.850 0.073 0.695 0.702 0.009 0.377 0.798 0.138 0.223 0.587 0.158 0.052
#>  [817] 0.968 0.121 0.363 0.184 0.667 0.519 0.868 0.546 0.578 0.168 0.242 0.829
#>  [829] 0.802 0.450 0.390 0.985 0.151 0.063 0.441 0.535 0.561 0.977 0.641 0.277
#>  [841] 0.440 0.729 0.854 0.846 0.927 0.611 0.051 0.255 0.090 0.477 0.500 0.730
#>  [853] 0.678 0.759 0.258 0.124 0.712 0.285 0.419 0.878 0.084 0.300 0.749 0.033
#>  [865] 0.790 0.751 0.517 0.247 0.873 0.156 0.408 0.092 0.550 0.127 0.456 0.949
#>  [877] 0.880 0.512 0.858 0.375 0.353 0.453 0.918 0.382 0.637 0.087 0.915 0.991
#>  [889] 0.599 0.197 0.019 0.429 0.966 0.835 0.649 0.762 0.443 0.426 0.056 0.261
#>  [901] 0.670 0.405 0.538 0.714 0.876 0.079 0.689 0.577 0.960 0.479 0.111 0.741
#>  [913] 0.249 0.227 0.139 0.047 0.313 0.930 0.853 0.086 0.297 0.923 0.037 0.974
#>  [925] 0.095 0.098 0.845 0.703 0.416 0.770 0.595 0.392 0.497 0.288 0.097 0.754
#>  [937] 0.014 0.653 0.605 0.423 0.656 0.706 0.581 0.180 0.820 0.350 0.506 0.343
#>  [949] 0.106 0.007 0.338 0.640 0.028 0.045 0.048 0.309 0.099 0.038 0.934 0.890
#>  [961] 0.523 0.917 0.015 0.307 0.797 0.763 0.773 0.371 0.973 0.777 0.159 0.631
#>  [973] 0.833 0.944 0.257 0.476 0.554 0.913 0.596 0.294 0.939 0.256 0.630 0.891
#>  [985] 0.979 0.818 0.355 0.273 0.320 0.742 0.896 0.491 0.422 0.394 0.154 0.170
#>  [997] 0.617 0.969 0.897 0.207
#> 
#> $angles
#> [1] 3.281769 4.206626
#> 
#> $ellipse
#> Vector (Vec3) object (n = 11):
#>               x         y         z
#>  [1,] 0.3255999 0.8935846 0.3090167
#>  [2,] 0.3671404 0.8776759 0.3080470
#>  [3,] 0.3992291 0.8708523 0.2867619
#>  [4,] 0.4098296 0.8762430 0.2534522
#>  [5,] 0.3948981 0.8917916 0.2208240
#>  [6,] 0.3599145 0.9110346 0.2011901
#>  [7,] 0.3181238 0.9262850 0.2019737
#>  [8,] 0.2856305 0.9320435 0.2229577
#>  [9,] 0.2750301 0.9266528 0.2562674
#> [10,] 0.2903661 0.9121693 0.2891967
#> [11,] 0.3255999 0.8935846 0.3090167
#> 

# Check how many vectors lie outside quantiles:
quantile(ce$pvalue.line, probs=c(0.00, 0.05, 0.25, 0.50, 0.75, 1.00))
#>      0%      5%     25%     50%     75%    100% 
#> 0.00000 0.04995 0.24975 0.49950 0.74925 0.99900 

# Hypothesis testing (reject if p-value < alpha):
ce$pvalue.line.FUN((Line(90, 0)))
#> [1] 0