Deformation Gradient Tensor

Deformation Gradient Tensor

Usage

defgrad_from_ratio(Rxy = 1, Ryz = 1)

defgrad_from_pair(p)

defgrad_from_vectors(v1, v2)

defgrad_from_axisangle(axis, angle)

defgrad_from_comp(
  xx = 1,
  xy = 0,
  xz = 0,
  yx = 0,
  yy = 1,
  yz = 0,
  zx = 0,
  zy = 0,
  zz = 1
)

Arguments

Rxy, Ryz

numeric. the XY and YZ strain ratio to create a strain tensor with axial stretches.Values must be greater than or equal to 1.

p

object of class pair

v1, v2

objects of class "spherical" or three-element vector. Deformation gradient results from the rotation around axis perpendicular to both vectors to rotate v1 to v2.

axis, angle

rotation axis and angle, axis can be an object of class "spherical" (incl. "line" and "plane") or a three-element vector. Angle in degrees when axis is a object of class "spherical", and radians otherwise.

xx, xy, xz, yx, yy, yz, zx, zy, zz

numeric. Directly specify components of the tensor. Identity matrix by default.

Value

3x3 matrix.

Examples

defgrad_from_ratio(2, 3)
#>          [,1]     [,2]      [,3]
#> [1,] 2.289428 0.000000 0.0000000
#> [2,] 0.000000 1.144714 0.0000000
#> [3,] 0.000000 0.000000 0.3815714
defgrad_from_axisangle(Line(120, 30), 45)
#>            [,1]       [,2]      [,3]
#> [1,]  0.7620243 -0.4486733 0.4669168
#> [2,]  0.2584335  0.8718592 0.4160212
#> [3,] -0.5937433 -0.1963513 0.7803301
defgrad_from_vectors(Line(120, 30), Line(210, 60))
#>             [,1]       [,2]       [,3]
#> [1,]  0.97419453 -0.1931138  0.1168422
#> [2,]  0.22291137  0.9044366 -0.3637374
#> [3,] -0.03543362  0.3803964  0.9241445
defgrad_from_pair(Pair(40, 20, 75, 16))
#>           [,1]       [,2]       [,3]
#> [1,] 0.2487928 -0.9331095 -0.2620026
#> [2,] 0.9285075  0.3060065 -0.2198463
#> [3,] 0.2756374 -0.1885752  0.9396926