Helper function for Eigenvalues and Eigenvectors of a Set of Vectors

Helper function for Eigenvalues and Eigenvectors of a Set of Vectors

Usage

.or_eigen_helper(x, scaled = FALSE)

Arguments

x

Object of class "Vec3", "Line" or "Plane"

scaled

logical. Whether the Eigenvectors should be scaled by the Eigenvalues (only effective if x is in Cartesian coordinates).