Ellipsoid tensor from principal stretches

Return diagonal tensor defined by magnitudes of principal stretches

Usage

ellipsoid_from_stretch(x = 1, y = 1, z = 1)

Arguments

x, y, z

numeric. Magnitudes of principal stretches

Value

object of class "ellipsoid"

Details

The eigenvalues \(\lambda\) of the deformation matrix are the quadratic forms of the principal stretches \(s\) (\(s = 1 + \epsilon = l/l_0\)).

See also

Examples

el <- ellipsoid_from_stretch(4, 3, 1)
principal_stretch(el)
#> S1 S2 S3 
#>  4  3  1 

lapply(seq.int(nrow(hossack1968)), function(i) {
  ellipsoid_from_stretch(hossack1968[i, 3], hossack1968[i, 2], hossack1968[i, 1])
})
#> [[1]]
#> Ellipsoid tensor
#>        [,1]   [,2] [,3]
#> [1,] 334.89   0.00    0
#> [2,]   0.00 123.21    0
#> [3,]   0.00   0.00    1
#> 
#> [[2]]
#> Ellipsoid tensor
#>      [,1]  [,2] [,3]
#> [1,] 1024  0.00    0
#> [2,]    0 17.64    0
#> [3,]    0  0.00    1
#> 
#> [[3]]
#> Ellipsoid tensor
#>      [,1] [,2] [,3]
#> [1,]  900 0.00    0
#> [2,]    0 5.29    0
#> [3,]    0 0.00    1
#> 
#> [[4]]
#> Ellipsoid tensor
#>        [,1] [,2] [,3]
#> [1,] 275.56    0    0
#> [2,]   0.00   16    0
#> [3,]   0.00    0    1
#> 
#> [[5]]
#> Ellipsoid tensor
#>         [,1]  [,2] [,3]
#> [1,] 2948.49  0.00    0
#> [2,]    0.00 56.25    0
#> [3,]    0.00  0.00    1
#> 
#> [[6]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 912.04  0.00    0
#> [2,]   0.00 54.76    0
#> [3,]   0.00  0.00    1
#> 
#> [[7]]
#> Ellipsoid tensor
#>      [,1] [,2] [,3]
#> [1,] 1024 0.00    0
#> [2,]    0 4.41    0
#> [3,]    0 0.00    1
#> 
#> [[8]]
#> Ellipsoid tensor
#>         [,1]   [,2] [,3]
#> [1,] 1592.01   0.00    0
#> [2,]    0.00 249.64    0
#> [3,]    0.00   0.00    1
#> 
#> [[9]]
#> Ellipsoid tensor
#>         [,1]  [,2] [,3]
#> [1,] 1288.81  0.00    0
#> [2,]    0.00 11.56    0
#> [3,]    0.00  0.00    1
#> 
#> [[10]]
#> Ellipsoid tensor
#>      [,1]  [,2] [,3]
#> [1,]  196  0.00    0
#> [2,]    0 23.04    0
#> [3,]    0  0.00    1
#> 
#> [[11]]
#> Ellipsoid tensor
#>         [,1]  [,2] [,3]
#> [1,] 1317.69  0.00    0
#> [2,]    0.00 17.64    0
#> [3,]    0.00  0.00    1
#> 
#> [[12]]
#> Ellipsoid tensor
#>         [,1]  [,2] [,3]
#> [1,] 1190.25  0.00    0
#> [2,]    0.00 67.24    0
#> [3,]    0.00  0.00    1
#> 
#> [[13]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 556.96  0.00    0
#> [2,]   0.00 20.25    0
#> [3,]   0.00  0.00    1
#> 
#> [[14]]
#> Ellipsoid tensor
#>         [,1]  [,2] [,3]
#> [1,] 1505.44  0.00    0
#> [2,]    0.00 10.24    0
#> [3,]    0.00  0.00    1
#> 
#> [[15]]
#> Ellipsoid tensor
#>         [,1] [,2] [,3]
#> [1,] 1361.61 0.00    0
#> [2,]    0.00 6.25    0
#> [3,]    0.00 0.00    1
#> 
#> [[16]]
#> Ellipsoid tensor
#>      [,1] [,2] [,3]
#> [1,]  625 0.00    0
#> [2,]    0 4.41    0
#> [3,]    0 0.00    1
#> 
#> [[17]]
#> Ellipsoid tensor
#>        [,1] [,2] [,3]
#> [1,] 590.49 0.00    0
#> [2,]   0.00 4.41    0
#> [3,]   0.00 0.00    1
#> 
#> [[18]]
#> Ellipsoid tensor
#>        [,1] [,2] [,3]
#> [1,] 148.84 0.00    0
#> [2,]   0.00 6.76    0
#> [3,]   0.00 0.00    1
#> 
#> [[19]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 453.69  0.00    0
#> [2,]   0.00 32.49    0
#> [3,]   0.00  0.00    1
#> 
#> [[20]]
#> Ellipsoid tensor
#>      [,1] [,2] [,3]
#> [1,]  676 0.00    0
#> [2,]    0 8.41    0
#> [3,]    0 0.00    1
#> 
#> [[21]]
#> Ellipsoid tensor
#>      [,1] [,2] [,3]
#> [1,]  625 0.00    0
#> [2,]    0 5.29    0
#> [3,]    0 0.00    1
#> 
#> [[22]]
#> Ellipsoid tensor
#>      [,1]  [,2] [,3]
#> [1,]  484  0.00    0
#> [2,]    0 34.81    0
#> [3,]    0  0.00    1
#> 
#> [[23]]
#> Ellipsoid tensor
#>        [,1]   [,2] [,3]
#> [1,] 357.21   0.00    0
#> [2,]   0.00 327.61    0
#> [3,]   0.00   0.00    1
#> 
#> [[24]]
#> Ellipsoid tensor
#>        [,1] [,2] [,3]
#> [1,] 392.04 0.00    0
#> [2,]   0.00 8.41    0
#> [3,]   0.00 0.00    1
#> 
#> [[25]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 262.44  0.00    0
#> [2,]   0.00 19.36    0
#> [3,]   0.00  0.00    1
#> 
#> [[26]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 492.84  0.00    0
#> [2,]   0.00 24.01    0
#> [3,]   0.00  0.00    1
#> 
#> [[27]]
#> Ellipsoid tensor
#>        [,1] [,2] [,3]
#> [1,] 237.16    0    0
#> [2,]   0.00    9    0
#> [3,]   0.00    0    1
#> 
#> [[28]]
#> Ellipsoid tensor
#>      [,1]  [,2] [,3]
#> [1,]  729  0.00    0
#> [2,]    0 19.36    0
#> [3,]    0  0.00    1
#> 
#> [[29]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 268.96  0.00    0
#> [2,]   0.00 14.44    0
#> [3,]   0.00  0.00    1
#> 
#> [[30]]
#> Ellipsoid tensor
#>        [,1] [,2] [,3]
#> [1,] 404.01 0.00    0
#> [2,]   0.00 7.29    0
#> [3,]   0.00 0.00    1
#> 
#> [[31]]
#> Ellipsoid tensor
#>         [,1]  [,2] [,3]
#> [1,] 2237.29  0.00    0
#> [2,]    0.00 10.89    0
#> [3,]    0.00  0.00    1
#> 
#> [[32]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 470.89  0.00    0
#> [2,]   0.00 12.25    0
#> [3,]   0.00  0.00    1
#> 
#> [[33]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 187.69  0.00    0
#> [2,]   0.00 37.21    0
#> [3,]   0.00  0.00    1
#> 
#> [[34]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 313.29  0.00    0
#> [2,]   0.00 47.61    0
#> [3,]   0.00  0.00    1
#> 
#> [[35]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 712.89  0.00    0
#> [2,]   0.00 60.84    0
#> [3,]   0.00  0.00    1
#> 
#> [[36]]
#> Ellipsoid tensor
#>        [,1] [,2] [,3]
#> [1,] 313.29    0    0
#> [2,]   0.00   16    0
#> [3,]   0.00    0    1
#> 
#> [[37]]
#> Ellipsoid tensor
#>      [,1] [,2] [,3]
#> [1,] 1681    0    0
#> [2,]    0   16    0
#> [3,]    0    0    1
#> 
#> [[38]]
#> Ellipsoid tensor
#>         [,1]  [,2] [,3]
#> [1,] 3708.81  0.00    0
#> [2,]    0.00 19.36    0
#> [3,]    0.00  0.00    1
#> 
#> [[39]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 424.36  0.00    0
#> [2,]   0.00 12.96    0
#> [3,]   0.00  0.00    1
#> 
#> [[40]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 610.09  0.00    0
#> [2,]   0.00 10.24    0
#> [3,]   0.00  0.00    1
#> 
#> [[41]]
#> Ellipsoid tensor
#>         [,1]   [,2] [,3]
#> [1,] 1183.36   0.00    0
#> [2,]    0.00 268.96    0
#> [3,]    0.00   0.00    1
#> 
#> [[42]]
#> Ellipsoid tensor
#>        [,1]   [,2] [,3]
#> [1,] 466.56   0.00    0
#> [2,]   0.00 184.96    0
#> [3,]   0.00   0.00    1
#> 
#> [[43]]
#> Ellipsoid tensor
#>         [,1]   [,2] [,3]
#> [1,] 1049.76   0.00    0
#> [2,]    0.00 309.76    0
#> [3,]    0.00   0.00    1
#> 
#> [[44]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 457.96  0.00    0
#> [2,]   0.00 32.49    0
#> [3,]   0.00  0.00    1
#> 
#> [[45]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 630.01  0.00    0
#> [2,]   0.00 20.25    0
#> [3,]   0.00  0.00    1
#> 
#> [[46]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 231.04  0.00    0
#> [2,]   0.00 43.56    0
#> [3,]   0.00  0.00    1
#> 
#> [[47]]
#> Ellipsoid tensor
#>         [,1]  [,2] [,3]
#> [1,] 1036.84  0.00    0
#> [2,]    0.00 24.01    0
#> [3,]    0.00  0.00    1
#> 
#> [[48]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 571.21  0.00    0
#> [2,]   0.00 33.64    0
#> [3,]   0.00  0.00    1
#> 
#> [[49]]
#> Ellipsoid tensor
#>      [,1]  [,2] [,3]
#> [1,]  256  0.00    0
#> [2,]    0 59.29    0
#> [3,]    0  0.00    1
#> 
#> [[50]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 789.61  0.00    0
#> [2,]   0.00 29.16    0
#> [3,]   0.00  0.00    1
#> 
#> [[51]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 432.64  0.00    0
#> [2,]   0.00 31.36    0
#> [3,]   0.00  0.00    1
#> 
#> [[52]]
#> Ellipsoid tensor
#>       [,1] [,2] [,3]
#> [1,] 77.44 0.00    0
#> [2,]  0.00 8.41    0
#> [3,]  0.00 0.00    1
#> 
#> [[53]]
#> Ellipsoid tensor
#>       [,1] [,2] [,3]
#> [1,] 16.81    0    0
#> [2,]  0.00   16    0
#> [3,]  0.00    0    1
#> 
#> [[54]]
#> Ellipsoid tensor
#>       [,1]  [,2] [,3]
#> [1,] 26.01  0.00    0
#> [2,]  0.00 17.64    0
#> [3,]  0.00  0.00    1
#> 
#> [[55]]
#> Ellipsoid tensor
#>       [,1]  [,2] [,3]
#> [1,] 50.41  0.00    0
#> [2,]  0.00 12.25    0
#> [3,]  0.00  0.00    1
#> 
#> [[56]]
#> Ellipsoid tensor
#>       [,1] [,2] [,3]
#> [1,] 39.69 0.00    0
#> [2,]  0.00 9.61    0
#> [3,]  0.00 0.00    1
#> 
#> [[57]]
#> Ellipsoid tensor
#>       [,1] [,2] [,3]
#> [1,] 33.64 0.00    0
#> [2,]  0.00 7.84    0
#> [3,]  0.00 0.00    1
#> 
#> [[58]]
#> Ellipsoid tensor
#>       [,1]  [,2] [,3]
#> [1,] 59.29  0.00    0
#> [2,]  0.00 14.44    0
#> [3,]  0.00  0.00    1
#> 
#> [[59]]
#> Ellipsoid tensor
#>       [,1] [,2] [,3]
#> [1,] 88.36    0    0
#> [2,]  0.00   16    0
#> [3,]  0.00    0    1
#> 
#> [[60]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 342.25  0.00    0
#> [2,]   0.00 34.81    0
#> [3,]   0.00  0.00    1
#> 
#> [[61]]
#> Ellipsoid tensor
#>       [,1]  [,2] [,3]
#> [1,] 46.24  0.00    0
#> [2,]  0.00 12.96    0
#> [3,]  0.00  0.00    1
#> 
#> [[62]]
#> Ellipsoid tensor
#>       [,1]  [,2] [,3]
#> [1,] 50.41  0.00    0
#> [2,]  0.00 15.21    0
#> [3,]  0.00  0.00    1
#> 
#> [[63]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 112.36  0.00    0
#> [2,]   0.00 20.25    0
#> [3,]   0.00  0.00    1
#> 
#> [[64]]
#> Ellipsoid tensor
#>      [,1] [,2] [,3]
#> [1,]   64    0    0
#> [2,]    0   16    0
#> [3,]    0    0    1
#> 
#> [[65]]
#> Ellipsoid tensor
#>       [,1]  [,2] [,3]
#> [1,] 51.84  0.00    0
#> [2,]  0.00 17.64    0
#> [3,]  0.00  0.00    1
#> 
#> [[66]]
#> Ellipsoid tensor
#>       [,1] [,2] [,3]
#> [1,] 98.01    0    0
#> [2,]  0.00   16    0
#> [3,]  0.00    0    1
#> 
#> [[67]]
#> Ellipsoid tensor
#>      [,1]  [,2] [,3]
#> [1,]   16  0.00    0
#> [2,]    0 12.25    0
#> [3,]    0  0.00    1
#> 
#> [[68]]
#> Ellipsoid tensor
#>       [,1] [,2] [,3]
#> [1,] 40.96 0.00    0
#> [2,]  0.00 7.29    0
#> [3,]  0.00 0.00    1
#> 
#> [[69]]
#> Ellipsoid tensor
#>       [,1]  [,2] [,3]
#> [1,] 67.24  0.00    0
#> [2,]  0.00 12.96    0
#> [3,]  0.00  0.00    1
#> 
#> [[70]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 118.81  0.00    0
#> [2,]   0.00 18.49    0
#> [3,]   0.00  0.00    1
#> 
#> [[71]]
#> Ellipsoid tensor
#>       [,1]  [,2] [,3]
#> [1,] 59.29  0.00    0
#> [2,]  0.00 14.44    0
#> [3,]  0.00  0.00    1
#> 
#> [[72]]
#> Ellipsoid tensor
#>       [,1]  [,2] [,3]
#> [1,] 88.36  0.00    0
#> [2,]  0.00 16.81    0
#> [3,]  0.00  0.00    1
#> 
#> [[73]]
#> Ellipsoid tensor
#>       [,1]  [,2] [,3]
#> [1,] 72.25  0.00    0
#> [2,]  0.00 12.96    0
#> [3,]  0.00  0.00    1
#> 
#> [[74]]
#> Ellipsoid tensor
#>       [,1]  [,2] [,3]
#> [1,] 72.25  0.00    0
#> [2,]  0.00 17.64    0
#> [3,]  0.00  0.00    1
#> 
#> [[75]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 163.84  0.00    0
#> [2,]   0.00 21.16    0
#> [3,]   0.00  0.00    1
#> 
#> [[76]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 158.76  0.00    0
#> [2,]   0.00 15.21    0
#> [3,]   0.00  0.00    1
#> 
#> [[77]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 182.25  0.00    0
#> [2,]   0.00 20.25    0
#> [3,]   0.00  0.00    1
#> 
#> [[78]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 204.49  0.00    0
#> [2,]   0.00 27.04    0
#> [3,]   0.00  0.00    1
#> 
#> [[79]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 262.44  0.00    0
#> [2,]   0.00 17.64    0
#> [3,]   0.00  0.00    1
#> 
#> [[80]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 201.64  0.00    0
#> [2,]   0.00 70.56    0
#> [3,]   0.00  0.00    1
#> 
#> [[81]]
#> Ellipsoid tensor
#>        [,1]  [,2] [,3]
#> [1,] 102.01  0.00    0
#> [2,]   0.00 26.01    0
#> [3,]   0.00  0.00    1
#>