Fabric intensity and shape indices

Fabric intensity and shape indices

Usage

fabric_indexes(x)

Arguments

x

numeric. Can be three element vector, three column array, or an object of class "line" or "plane"

Value

numeric vector containing the fabric shape and intensity indices:

P

Point (Vollmer 1990), range: (0, 1)

G

Girdle (Vollmer 1990), range: (0, 1)

R

Random (Vollmer 1990), range: (0, 1)

B

cylindricity (Vollmer 1990), range: (0, 1)

C

cylindricity or Fabric strength (Woodcock 1977), range: (0, Inf)

I

cylindricity or Fabric intensity (Lisle 1985), range: (0, 5)

Examples

set.seed(1)
mu <- Line(120, 50)
x <- rvmf(100, mu = mu, k = 1)
fabric_indexes(x)
#>          P          G          R          B          C          I 
#> 0.12518955 0.09964774 0.77516271 0.22483729 0.51720272 2.45026220