Fault deformation gradient tensor

Creates the fault deformation gradient tensor from displacement components. If the dip direction is know, the tensor will be rotated into the geographic reference frame.

Usage

fault_tensor(displacements, dip_direction = NULL)

Arguments

displacements

data.frame containing the strike slip, the heave, and the vertical throw of the fault's displacement components.

dip_direction

(optional) dip direction in degrees. If NULL, the fault tensor will be given in the fault displacement coordinates. Otherwise, the tensor will be in the geographic reference frame.

Details

x axis of tensor = heave, y = strike slip, z = vertical throw (positive for thrusting, negative for normal faulting)

Examples

if (FALSE) { # \dontrun{
fault_tensor(displacements = data.frame(strikeslip = 2, verticalthrow = -5, heave = 3), dip_direction = 0)
} # }