Fisher's statistics

Estimates concentration parameter, angular standard deviation, and confidence limit.

Usage

fisher_statistics(x, w = NULL, p = 0.05)

Arguments

x

numeric. Can be three element vector, three column array, or an object of class "line" or "plane"

w

numeric. Weights

p

numeric. Significance level (0.05 by default), corresponding to \(100 * (1-p) - 95\%\) confidence level.

Value

list, with

"k"

estimated concentration parameter \(\kappa\) for the von Mises-Fisher distribution

"csd"

estimated angular standard deviation enclosing 63% of the orientation data. Angle is in degrees if x is a spherical object, and raidan if otherwise.

"a95"

Confidence limit for given p. Angle is in degrees if x is a spherical object, and raidan if otherwise.

Examples

x <- rvmf(100, mu = Line(120, 50), k = 5)
fisher_statistics(x)
#> $k
#> [1] 3.894884
#> 
#> $csd
#> [1] 41.04287
#> 
#> $csd_2s
#> [1] 70.9383
#> 
#> $alpha
#> [1] 8.298471
#>