The Frechet (geodesic \(L^2\)) mean

An iterative algorithm for computing the Frechet mean, i.e. the vector that minimizes the Frechet variance.

Usage

geodesic_mean(x, ...)

geodesic_var(x, ...)

Source

geologyGeometry (J.R. Davis)

Arguments

x

object of class "Vec3", "Line", "Ray", "Plane", "Pair", or "Fault".

...

parameters passed to geodesic_meanvariance_ray() (if x is a Ray), geodesic_meanvariance_line() (if x is a Vec3, Line or Plane) or geodesic_mean_pair() (if x is a Pair or a Fault).

Value

geodesic_mean returns the mean vector as an object of class x. geodesic_var returns the variance as a numeric number.

References

Davis, J. R., & Titus, S. J. (2017). Modern methods of analysis for three-dimensional orientational data. Journal of Structural Geology, 96, 65–89. https://doi.org/10.1016/j.jsg.2017.01.002

See also

Examples

geodesic_mean(example_planes)
#> Plane object (n = 1):
#> dip_direction           dip 
#>      345.9172       75.3472 
geodesic_var(example_planes)
#> [1] 0.2656372