Calculates densities for fabric and strain data

Densities in hyperboloidal projections of geological fabric and finite strain data density calculations done on the unit hyperboloid (Vollmer, 2018). Options are given for equidistant (Elliott), equal-area, stereographic, orthographic, exponential, and radial projections, as polar azimuthal or cylindrical (cartesian, RfPhi-type) plots.

Usage

hypercontour(
  r,
  phi,
  proj = c("eqd", "eqa", "stg", "ort", "gno", "lin", "rdl", "rfp"),
  normalize = TRUE,
  rmax = NULL,
  kappa = 40,
  nnodes = 50L
)

Arguments

r

numeric. Aspect ratio of deformed object (long axis / short axis)

phi

numeric. Orientation of long axis of deformed object (in degrees)

proj

character. Projection, 'eqd' for equidistant (Elliot plot), 'eqa' for equal-area, 'stg' for stereographic, 'ort' for orthographic, 'gno' for gnomonic, 'lin' for exponential (linear R), 'rdl' for radial, 'rfp' for Rf/phi (cylindrical instead of polar).

normalize

logical.

rmax

maximum R value (if NULL, computed automatically)

kappa

smoothing parameter

nnodes

grid resolution. higher is more accurate but slower, 30 is good, but 50 is recommended for final plots, default = 50.

Value

list with:

x, y: vectors of the projection coordinates for the density grid

z: density matrix

points: projected data points

frame: plot frame (circle or square)

Details

The data must be in a comma delimited csv text file with one (R, phi) pair per line, where R = strain ratio (max/min), phi = orientation of long (max) axis from x. Contours are equally spaced over the probability density distribution.

References

Vollmer, F.W., 2018. Automatic contouring of geological fabric and finite strain data on the unit hyperboloid. Computers & Geosciences, https://doi.org/10.1016/j.cageo.2018.03.006

See also

Examples

data(ramsay)
out <- hypercontour(ramsay[, "R"], ramsay[, "phi"], proj = "rfp")
head(out$z)
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 114.3282 114.3282 114.3282 114.3282 114.3282 114.3282 114.3282 114.3282
#> [2,] 114.3282 114.3282 114.3282 114.3282 114.3282 114.3282 114.3282 114.3282
#> [3,] 114.3282 114.3282 114.3282 114.3282 114.3282 114.3282 114.3282 114.3282
#> [4,] 114.3282 114.3282 114.3282 114.3282 114.3282 114.3282 114.3282 114.3282
#> [5,] 114.3282 114.3282 114.3282 114.3282 114.3282 114.3282 114.3282 114.3282
#> [6,] 114.3282 114.3282 114.3282 114.3282 114.3282 114.3282 114.3282 114.3282
#>          [,9]    [,10]    [,11]    [,12]    [,13]    [,14]    [,15]    [,16]
#> [1,] 114.3282 114.3282 110.3759 90.39486 71.71766 55.57557 42.31459 31.79472
#> [2,] 114.3282 114.3282 109.8026 87.77160 68.14104 51.77714 38.72344 28.62373
#> [3,] 114.3282 114.3282 109.3024 85.53477 65.15324 48.66214 35.82725 26.10478
#> [4,] 114.3282 114.3282 108.8825 83.69222 62.73330 46.17674 33.54702 24.14481
#> [5,] 114.3282 114.3282 108.5486 82.24874 60.86130 44.27499 31.81850 22.67076
#> [6,] 114.3282 114.3282 108.3055 81.20715 59.52034 42.92046 30.59254 21.62821
#>         [,17]    [,18]     [,19]    [,20]    [,21]    [,22]    [,23]    [,24]
#> [1,] 23.65548 17.47228 12.838195 9.399737 6.867117 5.011499 3.656806 2.670043
#> [2,] 20.97854 15.28258 11.087781 8.024332 5.800334 4.192203 3.032234 2.196518
#> [3,] 18.88088 13.58759  9.747476 6.981120 4.997739 3.579927 2.567942 1.845867
#> [4,] 17.26551 12.29393  8.732227 6.195800 4.396479 3.122852 2.222102 1.584916
#> [5,] 16.05851 11.33230  7.980477 5.615835 3.953087 2.785899 1.966970 1.392093
#> [6,] 15.20611 10.65335  7.449226 5.205180 3.638230 2.545751 1.784362 1.253436
#>          [,25]     [,26]     [,27]     [,28]     [,29]     [,30]     [,31]
#> [1,] 1.9521005 1.4298573 1.0497537 0.7727642 0.5705568 0.4226101 0.3140789
#> [2,] 1.5944958 1.1605000 0.8471758 0.6205022 0.4560938 0.3364912 0.2491959
#> [3,] 1.3303243 0.9617137 0.6976078 0.5078804 0.3711665 0.2723172 0.2005832
#> [4,] 1.1336709 0.8135197 0.5858272 0.4234234 0.3072099 0.2237553 0.1636016
#> [5,] 0.9880052 0.7034094 0.5024745 0.3601987 0.2591365 0.1871050 0.1355821
#> [6,] 0.8827435 0.6234467 0.4416521 0.3138564 0.2237583 0.1600409 0.1148352
#>           [,32]      [,33]      [,34]      [,35]      [,36]      [,37]
#> [1,] 0.23422800 0.17529125 0.13164443 0.09920805 0.07501696 0.05691056
#> [2,] 0.18525429 0.13824398 0.10354882 0.07784291 0.05872303 0.04444710
#> [3,] 0.14832587 0.11010627 0.08204127 0.06135027 0.04603545 0.03465615
#> [4,] 0.12007565 0.08845769 0.06539963 0.04851867 0.03611295 0.02696240
#> [5,] 0.09859452 0.07194411 0.05267168 0.03868455 0.02849764 0.02105331
#> [6,] 0.08265803 0.05967923 0.04321583 0.03138271 0.02285115 0.01668149
#>           [,38]       [,39]       [,40]       [,41]       [,42]       [,43]
#> [1,] 0.04331016 0.033058597 0.025304957 0.019421296 0.014942568 0.011523099
#> [2,] 0.03374787 0.025700094 0.019625587 0.015025319 0.011530598 0.008867948
#> [3,] 0.02616950 0.019817539 0.015047166 0.011453086 0.008737099 0.006678920
#> [4,] 0.02018894 0.015158076 0.011409500 0.008607936 0.006508213 0.004930356
#> [5,] 0.01559546 0.011581596 0.008620997 0.006431222 0.004807360 0.003600229
#> [6,] 0.01220698 0.008952946 0.006580306 0.004846054 0.003575477 0.002642576
#>            [,44]       [,45]       [,46]        [,47]        [,48]        [,49]
#> [1,] 0.008904999 0.006895160 0.005348462 0.0041554596 0.0032333295 0.0025191895
#> [2,] 0.006833709 0.005275634 0.004079476 0.0031592009 0.0024497928 0.0019019611
#> [3,] 0.005115175 0.003924238 0.003015234 0.0023200216 0.0017873442 0.0013785223
#> [4,] 0.003741760 0.002844370 0.002165433 0.0016507896 0.0012600032 0.0009627936
#> [5,] 0.002700850 0.002029352 0.001527019 0.0011505596 0.0008679642 0.0006555065
#> [6,] 0.001956210 0.001450265 0.001076654 0.0008003071 0.0005955906 0.0004437218
#>             [,50]
#> [1,] 0.0019651470
#> [2,] 0.0014782219
#> [3,] 0.0010642846
#> [4,] 0.0007364260
#> [5,] 0.0004955555
#> [6,] 0.0003309096