Mean orientation of a set of pairs or faults

The Frechet (geodesic \(L^2\)) mean and variance of a pair of foliations and lineations

Usage

geodesic_mean_pair(x, group = NULL)

geodesic_var_pair(x, group = NULL)

Source

oriMeanVariance from geologyGeometry (J.R. Davis)

Arguments

x

object of class "Pair" or "Fault"

group

character. Symmetry group of x. One of "orthorhombic" (line-in-plane symmetry, e.g. foliation-lineations, cylindrical fold orientations, triaxial ellipsoid orientations, and earthquake focal mechanisms, and olivine), "triclinic" (ray-in-plane symmetry, e.g. faults with slip directions), "trigonal" (e.g. alpha-quartz), "hexagonal" (e.g. beta-quartz), or "trivial" (rotations). If NULL, the group will be automatically picked based on the class of x.

Value

object of class "Pair" or "Fault", respectively

References

Davis, J. R., & Titus, S. J. (2017). Modern methods of analysis for three-dimensional orientational data. Journal of Structural Geology, 96, 65–89. https://doi.org/10.1016/j.jsg.2017.01.002

See also

Examples

my_fault <- Fault(
  c("a" = 120, "b" = 120, "c" = 100),
  c(60, 60, 50),
  c(110, 25, 30),
  c(58, 9, 23),
  c(1, -1, 1)
)
geodesic_mean_pair(my_fault)
#> Fault object (n = 1):
#> dip_direction           dip       azimuth        plunge         sense 
#>     111.44744      57.51825      46.27697      33.40906       1.00000 
geodesic_var_pair(my_fault)
#> [1] 0.1945106