Orientation tensor

3D orientation tensor, which characterize data distribution using eigenvalue method. See (Watson 1966, Scheidegger 1965).

Usage

ortensor(x, norm = TRUE, w = NULL)

Arguments

x

numeric. Can be three element vector, three column array, or an object of class "line" or "plane"

norm

logical. Whether the tensor should be normalized or not.

w

numeric. weightings

Value

matrix

Details

The normalized orientation tensor is given as $$D = \frac{1}{n} (x_i, y_i, z_i) (x_i, y_i, z_i)^T$$ n = 1

See also

Examples

set.seed(1)
x <- rfb(100, mu = Line(120, 50), k = 1, A = diag(c(10, 0, 0)))
ortensor(x)
#>             [,1]        [,2]        [,3]
#> [1,]  0.11008014 -0.15160550 -0.07321833
#> [2,] -0.15160550  0.47368735  0.04359148
#> [3,] -0.07321833  0.04359148  0.41623251