Uncertainties of direction stress inversion after Yamaji and Sato (2006)

Bootstrap resampling to evaluate solution precision (Section 6). Yields B stress tensors from resampled datasets. The dispersion of these tensors on S^5 approximates the noise level of the data (Eq. 37).

Usage

slip_inversion_yamaji_sato_boot(x, weights = NULL, n_boot = 500L)

Arguments

x

object of class "Pair" or "Fault" with at least 4 rows.

weights

numeric. Weightings for the faults. Must have the same length as x

n_boot

integer. Number of bootstrap replicates

Value

list:

optimal

slip_inversion_yamaji_sato() result for the full dataset

thetas

length-B vector of angular stress distances from optimal

dispersion_deg

mean angular stress distance (Theta-bar); approximates the noise level p of the data (Fig. 8 of paper)

sd_deg

standard deviation of Theta values

D_bar

mean Orife-Lisle distance from optimal

DM_bar

mean Michael distance from optimal

See also

Examples

slip_inversion_yamaji_sato_boot(angelier1990$AVB)
#> $optimal
#> $optimal$stress_tensor
#>            [,1]       [,2]      [,3]
#> [1,] -0.5953283  0.1737441 0.3615258
#> [2,]  0.1737441 -0.4261257 0.2223252
#> [3,]  0.3615258  0.2223252 1.0214540
#> 
#> $optimal$principal_axes
#> Line object (n = 3):
#>          azimuth    plunge
#> sigma1  36.56594 74.346683
#> sigma2 246.92410 13.592506
#> sigma3 155.07122  7.616779
#> 
#> $optimal$principal_vals
#>     sigma1     sigma2     sigma3 
#>  1.1399309 -0.4105345 -0.7293963 
#> 
#> $optimal$principal_faults
#> Fault object (n = 2):
#>      dip_direction       dip  azimuth   plunge sense
#> [1,]      162.2618  68.95608 126.3346 64.58613     1
#> [2,]      146.7537 126.14052 354.0366 50.58977     1
#> 
#> $optimal$stress_shape
#> $optimal$stress_shape$R
#> [1] 0.8294243
#> 
#> $optimal$stress_shape$phi
#> [1] 0.1705757
#> 
#> $optimal$stress_shape$bott
#> [1] 1.205656
#> 
#> $optimal$stress_shape$A_phi
#> [1] 0.1705757
#> 
#> $optimal$stress_shape$type
#> [1] "N"
#> 
#> 
#> $optimal$tau_mean
#> [1] 0.6001158
#> 
#> $optimal$stress_components
#>             normal       shear slip_tendency dilatation_tendency
#>  [1,] -0.641267976  0.39462510   -0.61538251           0.9528556
#>  [2,] -0.455183853  0.64135736   -1.40900727           0.8533095
#>  [3,] -0.525560757  0.54188332   -1.03105742           0.8909578
#>  [4,] -0.428455701  0.67411940   -1.57337013           0.8390113
#>  [5,] -0.064363808  0.83837065  -13.02549799           0.6442396
#>  [6,] -0.166412315  0.85745223   -5.15257679           0.6988307
#>  [7,] -0.357348406  0.73957152   -2.06960912           0.8009723
#>  [8,] -0.375258185  0.72027866   -1.91942159           0.8105531
#>  [9,] -0.688606648  0.26006313   -0.37766572           0.9781795
#> [10,] -0.680413941  0.27816766   -0.40882122           0.9737968
#> [11,] -0.714240026 -0.06954651    0.09737134           0.9918921
#> [12,] -0.679445362  0.29957839   -0.44091609           0.9732786
#> [13,] -0.652785795  0.31832994   -0.48764839           0.9590170
#> [14,] -0.524116602  0.58156231   -1.10960483           0.8901852
#> [15,]  0.253590889  0.93304924    3.67934844           0.4741492
#> [16,] -0.659793682  0.18051805   -0.27359772           0.9627659
#> [17,] -0.542109362  0.51698469   -0.95365387           0.8998105
#> [18,] -0.506321094  0.60593591   -1.19674237           0.8806655
#> [19,] -0.326811101  0.76304867   -2.33483095           0.7846363
#> [20,] -0.285320602  0.78855966   -2.76376699           0.7624409
#> [21,]  0.198494647  0.93170746    4.69386696           0.5036230
#> [22,]  0.187381302  0.93445276    4.98690506           0.5095681
#> [23,]  0.240103463  0.93016098    3.87400070           0.4813643
#> [24,]  0.820112941  0.69976180    0.85325053           0.1710872
#> [25,] -0.245871474  0.81732539   -3.32419772           0.7413375
#> [26,] -0.299901771  0.77169302   -2.57315258           0.7702411
#> [27,] -0.470674957  0.62988042   -1.33824928           0.8615965
#> [28,] -0.665506152  0.32961084   -0.49527842           0.9658218
#> [29,] -0.619687003  0.40155461   -0.64799585           0.9413108
#> [30,] -0.410116610  0.62896249   -1.53361867           0.8292007
#> [31,] -0.246030052  0.80363642   -3.26641567           0.7414223
#> [32,] -0.001496547  0.88934457 -594.26454353           0.6106087
#> [33,] -0.725878252  0.04861919   -0.06697982           0.9981180
#> 
#> $optimal$misfit
#> $optimal$misfit$alpha
#>  [1]  1.1202493  2.0536177  4.1593140 15.7822083 17.7699860  5.7906986
#>  [7]  3.1403013 14.8381427  0.1566777  3.7763224 79.1502359  1.5535490
#> [13]  1.6879700 11.6802965 10.8548707 40.2293170  5.7417792  3.2054355
#> [19]  0.6667046  9.0672886  6.0849437  0.1002293  9.2220042  3.8248192
#> [25] 15.9914523  5.5236375  6.7580359 14.9662354 14.0075329 16.7214363
#> [31] 16.8812196  2.3298302 22.9927797
#> 
#> $optimal$misfit$alpha_mean
#> [1] 11.14634
#> 
#> $optimal$misfit$rup
#>  [1]  54.448617  26.125136  37.866444  32.831663  30.560405  10.100874
#>  [7]  15.454931  28.946922  69.970657  67.982628 101.817554  65.427104
#> [13]  63.267640  36.837994  21.105546  85.156850  41.040137  30.394856
#> [19]  11.940788  17.537911  13.369792   7.903400  18.234399  20.114063
#> [25]  27.605154  14.191574  29.061924  63.990275  56.144452  36.925869
#> [31]  29.182800   4.922226  94.857305
#> 
#> $optimal$misfit$rup_mean
#> [1] 38.34297
#> 
#> $optimal$misfit$quality
#>  [1] acceptable good       good       good       good       good      
#>  [7] good       good       acceptable acceptable poor       acceptable
#> [13] acceptable good       good       poor       good       good      
#> [19] good       good       good       good       good       good      
#> [25] good       good       good       acceptable acceptable good      
#> [31] good       good       poor      
#> Levels: good acceptable poor
#> 
#> $optimal$misfit$quality_summary
#>       n_good n_acceptable       n_poor 
#>           23            7            3 
#> 
#> $optimal$misfit$flipped
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE
#> [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [25] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 
#> 
#> $optimal$SHmax
#> [1] 65.07122
#> 
#> $optimal$alpha
#>  [1]  1.1202493  2.0536177  4.1593140 15.7822083 17.7699860  5.7906986
#>  [7]  3.1403013 14.8381427  0.1566777  3.7763224 79.1502359  1.5535490
#> [13]  1.6879700 11.6802965 10.8548707 40.2293170  5.7417792  3.2054355
#> [19]  0.6667046  9.0672886  6.0849437  0.1002293  9.2220042  3.8248192
#> [25] 15.9914523  5.5236375  6.7580359 14.9662354 14.0075329 16.7214363
#> [31] 16.8812196  2.3298302 22.9927797
#> 
#> $optimal$mean_alpha
#> [1] 11.14634
#> 
#> $optimal$y
#> [1] -0.4209607 -0.3013164  0.7222770  0.2223252  0.3615258  0.1737441
#> 
#> $optimal$method
#> [1] "yamaji_sato"
#> 
#> 
#> $thetas
#>   [1]  4.344583  3.754088 36.999349  8.723584  9.947542 14.259610 24.283674
#>   [8]  3.095480 14.406892  4.616729 22.577845 16.592530 31.051414  7.959735
#>  [15] 29.177293  7.959168 24.553593 10.010926 20.274598 22.001559 30.659134
#>  [22] 30.898720 13.242058  4.644392 19.194522  5.052112 10.115815 16.395104
#>  [29] 36.917260 16.840503 16.671686 36.760818 37.726935 14.010273 11.052836
#>  [36] 17.013868 14.383661 11.317130  5.947206 35.835169  6.116442  6.637019
#>  [43] 28.578622 28.770636  6.368244 18.899866 10.075880 14.590344 28.100247
#>  [50]  5.153194 19.030775  4.973766 17.424622 16.067667  7.096554  6.074921
#>  [57]  3.140345  4.474275 12.727923  7.717050 21.064124  5.449700 25.826352
#>  [64] 27.887226  6.371628  8.897701 10.605640 28.163460 11.979263  8.916013
#>  [71]  6.388163 11.324408 15.314233 32.626799 24.791209 26.762011 10.647867
#>  [78] 24.350911 14.788777  8.238265 11.985380 22.683176 41.592154  6.357784
#>  [85] 27.504469 11.469555 37.720935 33.340133 12.770251  6.506647 12.511778
#>  [92]  4.198343  9.540176 16.404207 21.542155 10.640832 17.233679 40.643333
#>  [99] 12.084063 15.071463 14.443671 14.050126 32.092526 14.303794 29.476495
#> [106] 27.283188 10.741807 10.628624 12.340498 26.527148 13.025924 34.259965
#> [113]  6.820377 15.192824 22.202478 18.670578  4.379268 11.127323 11.892425
#> [120]  5.393725  4.253859  5.504738 33.193754  7.025130 12.120942  6.032438
#> [127]  9.092518 48.517259  7.919585 29.067950 10.821235 41.243724 22.879992
#> [134]  9.709193 17.709403 15.154084 13.464886 27.382453 14.398899 36.554534
#> [141]  8.328822  8.099722 14.653977  9.475923 10.880337  6.209571 15.420987
#> [148]  8.873068  7.778758 42.603676 26.380755  3.743777  7.467844 40.854097
#> [155] 32.639891 24.703148 28.362457 12.773532  8.581071 23.315497  4.609639
#> [162] 22.925608  5.955917 15.898325  8.832500 26.402818  5.363027 24.262445
#> [169] 18.084040 14.040479 23.031930 12.842813 26.420561 29.876606 38.267509
#> [176] 12.354885 14.979256 11.086530 12.305652 13.467039 11.145887 27.440378
#> [183]  6.690000 14.252634  5.175019 39.856349 12.561211 23.585321 21.532436
#> [190] 40.829731  7.049263 10.955756 21.094175 15.961695 13.852616 14.746053
#> [197] 26.538880 17.421148 10.316215 17.295521 14.436692  4.847675  6.067253
#> [204] 10.734449 15.892577  6.152803 10.223205 34.000975  6.340683 31.056322
#> [211] 18.146859 14.970756  7.450121 17.728918 10.771737 31.156838 16.458479
#> [218] 12.398601 11.183721  3.811937  7.460448 30.283334 11.808723 14.376532
#> [225] 25.446346 21.060206 13.130908 24.161754 12.922666  4.248162  3.020542
#> [232] 23.112575 39.899672  4.669380  3.313435 26.259628 10.889524 12.067649
#> [239] 18.467180 16.938629 33.452004 41.417743  6.095734 25.319991 20.220405
#> [246] 34.734332  7.760090 35.420597 14.273172  5.108282 42.929023  7.109590
#> [253] 22.493050  5.553341 30.392610 34.112937 35.734054 16.302091  4.966035
#> [260] 11.491261 12.201068  6.576344 12.544285 27.351788 10.901192 29.389662
#> [267] 15.155743 16.722742 13.622069 20.402242 33.252144 17.485359 11.996596
#> [274] 23.210482  9.338061 16.542058 15.148900  7.579666 33.102247 25.186529
#> [281] 36.483171  4.701095 18.837549 52.523942 13.023247 11.143932  3.354125
#> [288]  8.477284  1.583700 11.198107 24.715229 26.591970  7.935399 30.365163
#> [295] 16.097398  3.587351 25.550177 34.342427 13.603014 11.412512 26.035058
#> [302]  5.600496  8.731350  8.963326  9.765301 13.824658  9.131057  6.004459
#> [309] 12.938195 16.536358  7.671212  4.184563 22.026818  4.812385  8.040549
#> [316] 12.468756 24.496854  7.523973 15.972237 10.300372  7.080863 15.112670
#> [323] 18.609761 13.481407 38.308820  6.160971 15.354957 23.870344 16.768076
#> [330]  8.714443 15.355599 14.702751  3.843755 20.381024 32.997884 31.980194
#> [337] 17.369707 12.807595  3.579229 27.476960 39.052321 21.540087 15.596650
#> [344] 29.184208 32.841026 12.814327  4.632515 12.928231  9.390866 30.099698
#> [351]  3.559747 31.173951 22.045758 11.819917 21.063296 10.786943 19.991673
#> [358] 11.443722 10.470749 19.906924 16.610595 10.958503  9.887706 46.892683
#> [365]  4.680271 40.032787  4.902557 15.624348 26.482020 11.588963  6.260066
#> [372] 16.417396  6.641089  8.697894 12.704541 12.876830 10.402289 10.575770
#> [379] 25.932943  9.890923 25.725787  7.060674 11.672252 23.053174 15.527289
#> [386]  6.627921 34.548451  8.539809 28.758389  8.712785  6.557322 48.722347
#> [393]  5.999800 19.754573 19.194836  4.949460  7.987562  4.117821 19.564633
#> [400] 10.035649 13.362102 23.013468 15.871452 17.911941  8.068836 35.894273
#> [407]  9.633969 16.081955 36.934854 15.530923  4.745507  4.709075 14.874997
#> [414] 14.294208 27.298433 30.385042  7.758605 12.200021 13.637047 36.775284
#> [421] 19.803833  9.120316 21.393067 31.579744 14.576134 10.122238 19.468026
#> [428]  5.904125  3.747353  7.849598 14.920032 13.838893 24.869640 10.241587
#> [435]  7.324656 17.044270 14.025004 32.999990  6.995289 33.161184 25.568590
#> [442] 23.573403 51.387399 14.892912  4.635802 12.205034 24.851682  4.658494
#> [449] 18.740603 18.326153 29.543760 22.533973 43.852849 33.000797  7.854483
#> [456] 48.510364 44.595430 13.082016 11.933144 36.797855 22.465928 11.583222
#> [463] 20.535421 11.954712  3.847619  8.462561  9.186247 13.435363 17.894313
#> [470]  8.960055 30.603741  6.278457 29.965307 25.712814  9.895632  5.395502
#> [477]  6.910037 50.296930  5.836175 16.047418 16.142127 36.821340 19.794604
#> [484] 16.256360  7.591128 23.451239 27.720056 26.276426 13.353860 28.272198
#> [491] 20.446226 34.194961 15.048190 25.712069 13.303219  8.264012 36.146649
#> [498] 22.881480 15.948114 10.687942
#> 
#> $dispersion_deg
#> [1] 17.26983
#> 
#> $sd_deg
#> [1] 10.64169
#> 
#> $D_bar
#> [1] 0.2987521
#> 
#> $DM_bar
#> [1] 0.06109465
#>