Sorting StraboSpot Projects

Sorts strabo objects in the order determined by one or more other objects, typically columns in the data data.frame. This will also sort the planar and linear objects.

Usage

# S3 method for class 'strabo'
sort_by(x, y, ...)

Arguments

x

object of class "strabo"

y

Variables to sort by. Typically a column in x$data data.frame. To sort by more than one column, take the form list(g1, g2, ...).

...

Additional arguments, typically passed on to order().

Value

A sorted version of x.

Examples

sort_by(strabo_prj, strabo_prj$data$dip_direction)
#> $data
#>                                        id dip_direction   dip strike plunge
#>                                    <char>         <num> <num>  <num>  <num>
#>   1: 8eafa098-d0f9-4f5a-bea1-32ecacfdf67e             0    42    270      6
#>   2: 201a5d08-a890-454d-8132-f795dc9e78c5             0    61    270     50
#>   3: 5db25631-768a-4dbc-991c-ecf14e24d938             2    41    272     11
#>   4: 79276f3e-87eb-4aae-807a-d188c5fd5b98             2    39    272     16
#>   5: 66c1d7bf-3f4d-4e2b-959b-174abbcbcc8a             3    81    273      4
#>  ---                                                                       
#> 343: e0500801-9bfc-487f-8cb3-5b4d536c7cb3           354    30    264      8
#> 344: e3695949-91f3-469c-9a88-5398b8687208           356    88    266      0
#> 345: d3777676-2f40-49d3-822f-ea188ba5f1ab           357    36    267     11
#> 346: 22d34b8d-9b19-4aa7-be0b-dbcec3244ac5           358    83    268     10
#> 347: e0c76bf2-823d-475d-afda-65aa8435407f           358    87    268     14
#>      trend associated planar_type quality unix_timestamp  notes
#>      <num>     <lgcl>      <char>  <char>          <num> <char>
#>   1:    83       TRUE        <NA>    <NA>   1.692891e+12   <NA>
#>   2:   312       TRUE        <NA>       5   1.693082e+12   <NA>
#>   3:    79       TRUE        <NA>    <NA>   1.692891e+12   <NA>
#>   4:    71       TRUE        <NA>    <NA>   1.692897e+12   <NA>
#>   5:    92       TRUE        <NA>       5   1.687873e+12   <NA>
#>  ---                                                           
#> 343:    69       TRUE        <NA>    <NA>   1.692897e+12   <NA>
#> 344:   266       TRUE        <NA>    <NA>   1.687792e+12   <NA>
#> 345:    71       TRUE        <NA>    <NA>   1.692893e+12   <NA>
#> 346:    87       TRUE        <NA>    <NA>   1.692645e+12   <NA>
#> 347:   269       TRUE        <NA>    <NA>   1.692645e+12   <NA>
#>      modified_timestamp           spot        spot_id feature_type
#>                   <num>         <char>         <char>       <char>
#>   1:                 NA 23-TS-Moss-139 16928905110927    foliation
#>   2:                 NA 23-TS-Moss-170 16930814935740    foliation
#>   3:                 NA 23-TS-Moss-139 16928905110927    foliation
#>   4:                 NA 23-TS-Moss-141 16928973747488    foliation
#>   5:                 NA  23-TS-Moss-51 16878724705561    foliation
#>  ---                                                              
#> 343:                 NA 23-TS-Moss-141 16928973747488    foliation
#> 344:                 NA  23-TS-Moss-40 16877910624414    foliation
#> 345:                 NA 23-TS-Moss-140 16928926111747    foliation
#> 346:                 NA 23-TS-Moss-107 16926438777524    foliation
#> 347:                 NA 23-TS-Moss-107 16926438777524    foliation
#>       foliation_type  label contact_type fault_or_sz_type bedding_type
#>               <char> <char>       <char>           <char>       <char>
#>   1:            <NA>   <NA>         <NA>             <NA>         <NA>
#>   2:     schistosity   <NA>         <NA>             <NA>         <NA>
#>   3:            <NA>   <NA>         <NA>             <NA>         <NA>
#>   4:            <NA>   <NA>         <NA>             <NA>         <NA>
#>   5:     schistosity   <NA>         <NA>             <NA>         <NA>
#>  ---                                                                  
#> 343:            <NA>   <NA>         <NA>             <NA>         <NA>
#> 344:        cleavage   <NA>         <NA>             <NA>         <NA>
#> 345:            <NA>   <NA>         <NA>             <NA>         <NA>
#> 346: slatey_cleavage   <NA>         <NA>             <NA>         <NA>
#> 347: slatey_cleavage   <NA>         <NA>             <NA>         <NA>
#>      vein_fill directional_indicators vein_type other_vein_fill
#>         <char>                 <char>    <char>          <char>
#>   1:      <NA>                   <NA>      <NA>            <NA>
#>   2:      <NA>                   <NA>      <NA>            <NA>
#>   3:      <NA>                   <NA>      <NA>            <NA>
#>   4:      <NA>                   <NA>      <NA>            <NA>
#>   5:      <NA>                   <NA>      <NA>            <NA>
#>  ---                                                           
#> 343:      <NA>                   <NA>      <NA>            <NA>
#> 344:      <NA>                   <NA>      <NA>            <NA>
#> 345:      <NA>                   <NA>      <NA>            <NA>
#> 346:      <NA>                   <NA>      <NA>            <NA>
#> 347:      <NA>                   <NA>      <NA>            <NA>
#>      foliation_defined_by movement other_feature fracture_type
#>                    <char>   <char>        <char>        <char>
#>   1:                 <NA>     <NA>          <NA>          <NA>
#>   2:                 <NA>     <NA>          <NA>          <NA>
#>   3:                 <NA>     <NA>          <NA>          <NA>
#>   4:                 <NA>     <NA>          <NA>          <NA>
#>   5:                 <NA>     <NA>          <NA>          <NA>
#>  ---                                                          
#> 343:                 <NA>     <NA>          <NA>          <NA>
#> 344:                 <NA>     <NA>          <NA>          <NA>
#> 345:                 <NA>     <NA>          <NA>          <NA>
#> 346:                 <NA>     <NA>          <NA>          <NA>
#> 347:                 <NA>     <NA>          <NA>          <NA>
#>      movement_justification        linear_type linear_quality linear_notes
#>                      <char>             <char>         <char>       <char>
#>   1:                   <NA> linear_orientation           <NA>         <NA>
#>   2:                   <NA> linear_orientation              5         <NA>
#>   3:                   <NA> linear_orientation           <NA>         <NA>
#>   4:                   <NA> linear_orientation           <NA>         <NA>
#>   5:                   <NA> linear_orientation              5         <NA>
#>  ---                                                                      
#> 343:                   <NA> linear_orientation           <NA>         <NA>
#> 344:                   <NA> linear_orientation           <NA>         <NA>
#> 345:                   <NA> linear_orientation           <NA>         <NA>
#> 346:                   <NA> linear_orientation           <NA>         <NA>
#> 347:                   <NA> linear_orientation           <NA>         <NA>
#>      linear_feature_type linear_defined_by linear_label               type
#>                   <char>            <char>       <char>             <char>
#>   1:          stretching              <NA>         <NA> planar_orientation
#>   2:          stretching               Hbl         <NA> planar_orientation
#>   3:          stretching              <NA>         <NA> planar_orientation
#>   4:          stretching              <NA>         <NA> planar_orientation
#>   5:          stretching               Hbl         <NA> planar_orientation
#>  ---                                                                      
#> 343:          stretching              <NA>         <NA> planar_orientation
#> 344:          stretching              <NA>         <NA> planar_orientation
#> 345:          stretching              <NA>         <NA> planar_orientation
#> 346:          stretching              <NA>         <NA> planar_orientation
#> 347:          stretching              <NA>         <NA> planar_orientation
#>      linear_unix_timestamp                            linear_id defined_by
#>                      <num>                               <char>     <char>
#>   1:          1.692891e+12 045ad2b9-5a85-4228-8273-55484b2ec6ce       <NA>
#>   2:          1.693082e+12 94bc8ef4-a758-4722-8ce5-236ba24ff444       <NA>
#>   3:          1.692891e+12 52e54e60-237e-4c9a-8873-9859cc445f97       <NA>
#>   4:          1.692897e+12 73357466-3526-4475-877a-4af76afcc696       <NA>
#>   5:          1.687873e+12 ff1ba044-0409-4db5-98bb-664a6604bac7       <NA>
#>  ---                                                                      
#> 343:          1.692897e+12 1df00205-dbda-4c79-b58a-c0db48e32abf       <NA>
#> 344:          1.687792e+12 4fec1a8e-a8f2-4a3f-9e8f-a1bd32179098       <NA>
#> 345:          1.692893e+12 b3203ffb-bab9-446a-9a8c-cff9dd003123       <NA>
#> 346:          1.692645e+12 52a946c8-6044-47d1-95c8-843f4e86b57d       <NA>
#> 347:          1.692645e+12 8c7944cc-10fe-4c11-ae01-4b31554e3d39       <NA>
#>      linear_other_feature
#>                    <char>
#>   1:                 <NA>
#>   2:                 <NA>
#>   3:                 <NA>
#>   4:                 <NA>
#>   5:                 <NA>
#>  ---                     
#> 343:                 <NA>
#> 344:                 <NA>
#> 345:                 <NA>
#> 346:                 <NA>
#> 347:                 <NA>
#> 
#> $planar
#> Plane object (n = 347):
#>        dip_direction dip
#>   [1,]             0  42
#>   [2,]             0  61
#>   [3,]             2  41
#>   [4,]             2  39
#>   [5,]             3  81
#>   [6,]             4  38
#>   [7,]             4  72
#>   [8,]             5  32
#>   [9,]             5  58
#>  [10,]             6  42
#>  [11,]             6  48
#>  [12,]             8  59
#>  [13,]             9  73
#>  [14,]            10  36
#>  [15,]            12  22
#>  [16,]            14  51
#>  [17,]            23  35
#>  [18,]            34  64
#>  [19,]            50  60
#>  [20,]            62  40
#>  [21,]            86  38
#>  [22,]            89  57
#>  [23,]           106  73
#>  [24,]           108  65
#>  [25,]           108  65
#>  [26,]           109  67
#>  [27,]           111  70
#>  [28,]           111  74
#>  [29,]           111  74
#>  [30,]           113  83
#>  [31,]           113  83
#>  [32,]           113  75
#>  [33,]           113  75
#>  [34,]           115  68
#>  [35,]           115  68
#>  [36,]           117  63
#>  [37,]           117  63
#>  [38,]           117  69
#>  [39,]           118  80
#>  [40,]           119  56
#>  [41,]           126  73
#>  [42,]           127  74
#>  [43,]           128  80
#>  [44,]           128  80
#>  [45,]           128  85
#>  [46,]           129  66
#>  [47,]           129  66
#>  [48,]           129  69
#>  [49,]           129  69
#>  [50,]           130  89
#>  [51,]           132  74
#>  [52,]           134  73
#>  [53,]           134  73
#>  [54,]           135  85
#>  [55,]           137  78
#>  [56,]           137  78
#>  [57,]           137  56
#>  [58,]           137  56
#>  [59,]           138  84
#>  [60,]           139  73
#>  [61,]           139  51
#>  [62,]           140  77
#>  [63,]           140  83
#>  [64,]           141  75
#>  [65,]           141  75
#>  [66,]           141  88
#>  [67,]           141  68
#>  [68,]           142  79
#>  [69,]           142  79
#>  [70,]           142  80
#>  [71,]           142  80
#>  [72,]           142  65
#>  [73,]           143  79
#>  [74,]           144  87
#>  [75,]           144  87
#>  [76,]           145  84
#>  [77,]           145  49
#>  [78,]           145  86
#>  [79,]           146  84
#>  [80,]           146  84
#>  [81,]           146  88
#>  [82,]           146  89
#>  [83,]           146  79
#>  [84,]           146  63
#>  [85,]           148  79
#>  [86,]           148  86
#>  [87,]           148  42
#>  [88,]           149  89
#>  [89,]           150  81
#>  [90,]           150  74
#>  [91,]           150  46
#>  [92,]           150  48
#>  [93,]           151  76
#>  [94,]           151  78
#>  [95,]           151  78
#>  [96,]           151  84
#>  [97,]           152  47
#>  [98,]           152  43
#>  [99,]           153  88
#> [100,]           153  88
#> [101,]           154  89
#> [102,]           154  70
#> [103,]           155  84
#> [104,]           156  88
#> [105,]           156  85
#> [106,]           157  69
#> [107,]           157  82
#> [108,]           158  77
#> [109,]           158  77
#> [110,]           158  88
#> [111,]           158  51
#> [112,]           158  90
#> [113,]           159  87
#> [114,]           159  87
#> [115,]           159  79
#> [116,]           159  86
#> [117,]           160  85
#> [118,]           160  55
#> [119,]           161  80
#> [120,]           161  77
#> [121,]           161  87
#> [122,]           161  84
#> [123,]           161  72
#> [124,]           161  88
#> [125,]           162  84
#> [126,]           162  76
#> [127,]           162  88
#> [128,]           162  85
#> [129,]           163  79
#> [130,]           163  85
#> [131,]           163  64
#> [132,]           164  73
#> [133,]           164  85
#> [134,]           164  87
#> [135,]           165  78
#> [136,]           165  75
#> [137,]           165  81
#> [138,]           165  89
#> [139,]           166  81
#> [140,]           166  85
#> [141,]           166  75
#> [142,]           166  89
#> [143,]           167  89
#> [144,]           167  68
#> [145,]           167  30
#> [146,]           168  68
#> [147,]           169  89
#> [148,]           169  77
#> [149,]           169  67
#> [150,]           170  81
#> [151,]           171  88
#> [152,]           173  89
#> [153,]           178  78
#> [154,]           180  87
#> [155,]           186  83
#> [156,]           186  78
#> [157,]           187  35
#> [158,]           190  33
#> [159,]           191  47
#> [160,]           191  84
#> [161,]           192  76
#> [162,]           193  79
#> [163,]           194  76
#> [164,]           194  32
#> [165,]           195  82
#> [166,]           195  36
#> [167,]           197  59
#> [168,]           210  77
#> [169,]           285  86
#> [170,]           286  68
#> [171,]           295  77
#> [172,]           297  74
#> [173,]           297  40
#> [174,]           299  48
#> [175,]           299  70
#> [176,]           304  71
#> [177,]           305  89
#> [178,]           306  77
#> [179,]           306  77
#> [180,]           306  59
#> [181,]           309  85
#> [182,]           311  78
#> [183,]           311  72
#> [184,]           311  46
#> [185,]           313  82
#> [186,]           313  81
#> [187,]           313  68
#> [188,]           314  88
#> [189,]           314  73
#> [190,]           315  79
#> [191,]           316  71
#> [192,]           316  85
#> [193,]           318  87
#> [194,]           318  73
#> [195,]           319  89
#> [196,]           319  71
#> [197,]           320  74
#> [198,]           320  68
#> [199,]           320  82
#> [200,]           321  47
#> [201,]           322  74
#> [202,]           322  54
#> [203,]           323  75
#> [204,]           323  77
#> [205,]           323  73
#> [206,]           324  81
#> [207,]           325  66
#> [208,]           325  82
#> [209,]           326  80
#> [210,]           326  88
#> [211,]           326  26
#> [212,]           327  89
#> [213,]           327  89
#> [214,]           327  83
#> [215,]           327  82
#> [216,]           328  89
#> [217,]           328  88
#> [218,]           328  37
#> [219,]           329  83
#> [220,]           329  83
#> [221,]           329  66
#> [222,]           329  83
#> [223,]           329  80
#> [224,]           329  66
#> [225,]           329  73
#> [226,]           330  82
#> [227,]           330  78
#> [228,]           330  76
#> [229,]           330  87
#> [230,]           330  84
#> [231,]           330  89
#> [232,]           330  89
#> [233,]           331  70
#> [234,]           331  87
#> [235,]           331  82
#> [236,]           331  70
#> [237,]           331  66
#> [238,]           332  87
#> [239,]           332  87
#> [240,]           332  74
#> [241,]           332  78
#> [242,]           332  84
#> [243,]           332  84
#> [244,]           332  88
#> [245,]           333  85
#> [246,]           333  75
#> [247,]           333  81
#> [248,]           333  81
#> [249,]           334  81
#> [250,]           334  87
#> [251,]           334  87
#> [252,]           334  87
#> [253,]           334  87
#> [254,]           334  81
#> [255,]           334  78
#> [256,]           334  78
#> [257,]           334  55
#> [258,]           335  81
#> [259,]           335  79
#> [260,]           335  84
#> [261,]           336  74
#> [262,]           336  84
#> [263,]           336  85
#> [264,]           336  85
#> [265,]           336  88
#> [266,]           336  72
#> [267,]           337  89
#> [268,]           337  72
#> [269,]           337  77
#> [270,]           337  89
#> [271,]           337  83
#> [272,]           337  81
#> [273,]           337  74
#> [274,]           338  85
#> [275,]           338  84
#> [276,]           338  87
#> [277,]           338  77
#> [278,]           338  58
#> [279,]           338  58
#> [280,]           339  83
#> [281,]           339  78
#> [282,]           339  85
#> [283,]           339  86
#> [284,]           339  60
#> [285,]           340  85
#> [286,]           340  84
#> [287,]           341  87
#> [288,]           341  62
#> [289,]           341  80
#> [290,]           341  81
#> [291,]           341  89
#> [292,]           341  73
#> [293,]           342  89
#> [294,]           342  86
#> [295,]           342  86
#> [296,]           342  78
#> [297,]           342  57
#> [298,]           343  84
#> [299,]           343  74
#> [300,]           343  74
#> [301,]           344  88
#> [302,]           344  64
#> [303,]           344  52
#> [304,]           344  76
#> [305,]           345  81
#> [306,]           345  50
#> [307,]           345  71
#> [308,]           345  74
#> [309,]           345  71
#> [310,]           346  82
#> [311,]           346  82
#> [312,]           346  85
#> [313,]           346  60
#> [314,]           346  65
#> [315,]           346  46
#> [316,]           347  89
#> [317,]           347  82
#> [318,]           347  88
#> [319,]           348  86
#> [320,]           348  82
#> [321,]           348  41
#> [322,]           348  52
#> [323,]           348  50
#> [324,]           348  86
#> [325,]           348  88
#> [326,]           349  86
#> [327,]           349  85
#> [328,]           349  41
#> [329,]           349  74
#> [330,]           350  84
#> [331,]           350  59
#> [332,]           350  29
#> [333,]           350  74
#> [334,]           351  88
#> [335,]           352  89
#> [336,]           352  53
#> [337,]           352  36
#> [338,]           352  42
#> [339,]           353  89
#> [340,]           353  82
#> [341,]           353  35
#> [342,]           353  60
#> [343,]           354  30
#> [344,]           356  88
#> [345,]           357  36
#> [346,]           358  83
#> [347,]           358  87
#> 
#> $linear
#> Line object (n = 347):
#>        azimuth plunge
#>   [1,]      83      6
#>   [2,]     312     50
#>   [3,]      79     11
#>   [4,]      71     16
#>   [5,]      92      4
#>   [6,]     278      3
#>   [7,]      65     57
#>   [8,]     347     31
#>   [9,]      81     22
#>  [10,]     291     13
#>  [11,]      79     17
#>  [12,]      86     20
#>  [13,]      93     17
#>  [14,]      61     25
#>  [15,]      31     20
#>  [16,]     293     10
#>  [17,]      65     28
#>  [18,]     111     25
#>  [19,]     139      2
#>  [20,]      49     40
#>  [21,]      94     37
#>  [22,]      73     56
#>  [23,]      17      1
#>  [24,]      19      3
#>  [25,]      19      3
#>  [26,]     183     35
#>  [27,]      22      3
#>  [28,]     201      0
#>  [29,]     201      0
#>  [30,]      24      7
#>  [31,]      24      7
#>  [32,]      26     11
#>  [33,]      26     11
#>  [34,]      29     10
#>  [35,]      29     10
#>  [36,]      31      7
#>  [37,]      31      7
#>  [38,]      28      2
#>  [39,]      29      7
#>  [40,]      51     29
#>  [41,]      49     36
#>  [42,]      53     46
#>  [43,]      39      3
#>  [44,]      39      3
#>  [45,]      39     10
#>  [46,]     213     13
#>  [47,]     213     13
#>  [48,]     213     15
#>  [49,]     213     15
#>  [50,]      40     19
#>  [51,]     163     71
#>  [52,]     217     23
#>  [53,]     217     23
#>  [54,]      48     33
#>  [55,]      47      3
#>  [56,]      47      3
#>  [57,]      52      7
#>  [58,]      52      7
#>  [59,]      48      4
#>  [60,]      72     53
#>  [61,]      63     17
#>  [62,]      55     22
#>  [63,]     230      2
#>  [64,]      52      4
#>  [65,]      52      4
#>  [66,]     230     44
#>  [67,]     225     13
#>  [68,]     220     48
#>  [69,]     224     37
#>  [70,]      55     16
#>  [71,]      55     16
#>  [72,]     194     53
#>  [73,]      58     24
#>  [74,]      55     12
#>  [75,]      55     12
#>  [76,]     235      2
#>  [77,]      67     13
#>  [78,]      55      9
#>  [79,]      57      7
#>  [80,]      57      7
#>  [81,]      57     19
#>  [82,]     236     19
#>  [83,]      57      5
#>  [84,]      64     15
#>  [85,]     228     44
#>  [86,]      60     35
#>  [87,]      74     14
#>  [88,]     239     14
#>  [89,]      66     33
#>  [90,]      66     21
#>  [91,]      80     19
#>  [92,]      75     16
#>  [93,]     208     66
#>  [94,]      62      4
#>  [95,]      62      4
#>  [96,]      64     29
#>  [97,]     241      1
#>  [98,]      81     16
#>  [99,]      63     12
#> [100,]      63     12
#> [101,]     244      4
#> [102,]      79     35
#> [103,]      67     21
#> [104,]      67     22
#> [105,]      67      9
#> [106,]      70      7
#> [107,]      67      1
#> [108,]     247      2
#> [109,]     247      2
#> [110,]      69     38
#> [111,]      76     10
#> [112,]     248     15
#> [113,]      70      6
#> [114,]      70     22
#> [115,]      71      8
#> [116,]      70     11
#> [117,]      73     27
#> [118,]      95     33
#> [119,]      74     20
#> [120,]      74     15
#> [121,]      71     10
#> [122,]      74     26
#> [123,]      76     14
#> [124,]      71      4
#> [125,]      72      1
#> [126,]     252      0
#> [127,]      72     16
#> [128,]      77     48
#> [129,]      78     25
#> [130,]      75     19
#> [131,]      81     15
#> [132,]      92     46
#> [133,]      76     24
#> [134,]     254     12
#> [135,]     199     76
#> [136,]      81     20
#> [137,]     255      1
#> [138,]     255     13
#> [139,]      76      4
#> [140,]      77     10
#> [141,]      84     28
#> [142,]      76      4
#> [143,]      77      5
#> [144,]      87     24
#> [145,]     174     30
#> [146,]     111     53
#> [147,]      79      0
#> [148,]      83     16
#> [149,]      86     18
#> [150,]      82     15
#> [151,]     261      0
#> [152,]      83      8
#> [153,]     266     10
#> [154,]      91     27
#> [155,]      96      4
#> [156,]      98      6
#> [157,]     151     30
#> [158,]     261     12
#> [159,]     227     41
#> [160,]     103     19
#> [161,]     281      4
#> [162,]     104      7
#> [163,]     105      4
#> [164,]     200     31
#> [165,]     106      2
#> [166,]     197     36
#> [167,]     124     26
#> [168,]     125     22
#> [169,]      14     17
#> [170,]     198      5
#> [171,]     206      4
#> [172,]      20     24
#> [173,]     222     12
#> [174,]     252     37
#> [175,]     213     12
#> [176,]      31      8
#> [177,]      35     73
#> [178,]      34      7
#> [179,]      34      7
#> [180,]     218      2
#> [181,]      38      7
#> [182,]      29     43
#> [183,]      41      0
#> [184,]     228      7
#> [185,]      37     38
#> [186,]     223      0
#> [187,]      35     17
#> [188,]      43     22
#> [189,]      44      1
#> [190,]     227     10
#> [191,]     226      0
#> [192,]     226      0
#> [193,]     228      1
#> [194,]      47      4
#> [195,]      49     10
#> [196,]      41     23
#> [197,]     231      6
#> [198,]     236     13
#> [199,]     232     14
#> [200,]     268     33
#> [201,]      48     15
#> [202,]      16     38
#> [203,]      48     17
#> [204,]      50     15
#> [205,]      36     45
#> [206,]     234      0
#> [207,]     236      2
#> [208,]     237     12
#> [209,]      51     30
#> [210,]     237     16
#> [211,]      54      1
#> [212,]      56      4
#> [213,]      56      4
#> [214,]     237      1
#> [215,]      57      0
#> [216,]      58      4
#> [217,]      57     21
#> [218,]      51      5
#> [219,]     239      0
#> [220,]     239      0
#> [221,]      58      1
#> [222,]     240     10
#> [223,]     240      2
#> [224,]     245     11
#> [225,]      49     30
#> [226,]      60      0
#> [227,]      57     15
#> [228,]     245     21
#> [229,]     240     10
#> [230,]     241     12
#> [231,]     240     14
#> [232,]     240     35
#> [233,]      58     10
#> [234,]      60      9
#> [235,]      59     16
#> [236,]     244      8
#> [237,]      54     15
#> [238,]      62      8
#> [239,]      62      8
#> [240,]      56     21
#> [241,]      60     13
#> [242,]      61     11
#> [243,]     245     31
#> [244,]      62      8
#> [245,]     243      8
#> [246,]      60     11
#> [247,]     249     30
#> [248,]      63      3
#> [249,]     246     14
#> [250,]      64     10
#> [251,]      64     10
#> [252,]      64      5
#> [253,]      64      5
#> [254,]     252     46
#> [255,]     245      3
#> [256,]      39     64
#> [257,]      53     16
#> [258,]      62     18
#> [259,]      63     13
#> [260,]     248     30
#> [261,]     285     66
#> [262,]      63     23
#> [263,]      65      8
#> [264,]     248     22
#> [265,]     247     38
#> [266,]      61     13
#> [267,]      67      4
#> [268,]      62     16
#> [269,]      62     24
#> [270,]      67     16
#> [271,]     250     25
#> [272,]      66      1
#> [273,]      65      6
#> [274,]      61     55
#> [275,]      65     23
#> [276,]     248     13
#> [277,]      67      4
#> [278,]      61     11
#> [279,]     255     10
#> [280,]      66     23
#> [281,]      67      8
#> [282,]      67     17
#> [283,]     253     39
#> [284,]     252      5
#> [285,]      69     13
#> [286,]      64     46
#> [287,]      71      3
#> [288,]      62     16
#> [289,]      70      8
#> [290,]     251      5
#> [291,]      70     25
#> [292,]      57     39
#> [293,]      71     12
#> [294,]      72      1
#> [295,]      71     12
#> [296,]     253      7
#> [297,]      61     18
#> [298,]      70     22
#> [299,]     254      3
#> [300,]      72      3
#> [301,]      74     23
#> [302,]      61     25
#> [303,]      67      9
#> [304,]     254      0
#> [305,]      73      9
#> [306,]      63     13
#> [307,]     258      9
#> [308,]     256      1
#> [309,]     261     17
#> [310,]      70     38
#> [311,]      75     10
#> [312,]      75     15
#> [313,]      71      9
#> [314,]      68     16
#> [315,]      52     22
#> [316,]      77     16
#> [317,]      76      4
#> [318,]     258      4
#> [319,]      76     23
#> [320,]      75     17
#> [321,]      68      8
#> [322,]      61     20
#> [323,]      60     20
#> [324,]      78      0
#> [325,]      78      9
#> [326,]     259      0
#> [327,]      76     32
#> [328,]      72      5
#> [329,]      77      5
#> [330,]      76     31
#> [331,]      66     22
#> [332,]      67      7
#> [333,]      74     21
#> [334,]     261      6
#> [335,]      82     17
#> [336,]      73     11
#> [337,]      64     12
#> [338,]      68     12
#> [339,]      83     17
#> [340,]      72     54
#> [341,]      64     13
#> [342,]      71     20
#> [343,]      69      8
#> [344,]     266      0
#> [345,]      71     11
#> [346,]      87     10
#> [347,]     269     14
#> 
#> attr(,"class")
#> [1] "list"   "strabo"