Import orientation data from StraboSpot

Reads the XLS format export of field book data, the JSON project file, or the txt export of StraboMobile data from strabospot.org/my_data and creates a list with the metadata, and the line or plane orientations.

Usage

read_strabo_xls(file, tag_cols = FALSE, sf = TRUE)

read_strabo_mobile(file, sf = TRUE)

read_strabo_JSON(file, sf = TRUE)

Arguments

file

the name of the file which the data are to be read from.

tag_cols

logical. Whether the Tag columns should be summarized in a single column (may lead to duplicate rows).

sf

logical. Whether the output should be a spatial "sf" object using the Longitude and Latitude columns.

Value

list containing the following objects:

data

"tbl_df" object or "sf" if sf == TRUE. Metadata.

tags

"tbl_df" object. Tags and their descriptsions.

planar

Plane elements. Same row IDs as in data.

linear

Line elements. Same row IDs as in data.

Examples

if (FALSE) { # \dontrun{
file <- "E:/Lakehead/Field Work/StraboSpot_Output_11_01_2022_TS.xlsx"
dt <- read_strabo_xls(file)
stereoplot()
stereo_greatcircle(dt$planar, col = "lightgrey")
stereo_point(dt$linear)

read_strabo_mobile("C:/Users/tobis/Downloads/StraboSpot_Search_06_16_2023.txt")

read_strabo_JSON("G:/My Drive/Moss_Lake/data.json")
} # }