Principal stretches, strain and shape parameters based on the orientation tensor.

Principal stretches, strain and shape parameters based on the orientation tensor.

Usage

principal_stretch(x)

principal_strain(x)

or_shape_params(x)

Arguments

x

numeric. Can be three element vector, three column array, or an object of class "line" or "plane"

Examples

set.seed(1)
mu <- Line(120, 50)
x <- rvmf(100, mu = mu, k = 20)
principal_stretch(x)
#>        S1        S2        S3 
#> 0.9511749 0.2220799 0.2143520 
principal_strain(x)
#>          e1          e2          e3 
#> -0.05005729 -1.50471812 -1.54013586 
or_shape_params(x)
#> $stretch_ratios
#>      Rxy      Ryz      Rxz 
#> 4.283031 1.036052 4.437444 
#> 
#> $strain_ratios
#>        e12        e13        e23 
#> 1.45466083 1.49007857 0.03541774 
#> 
#> $Vollmer
#>          P          G          R          B 
#> 0.85541428 0.00674541 0.13784031 0.86215969 
#> 
#> $Flinn
#> intensity  symmetry 
#>  3.283228 91.062705 
#> 
#> $Ramsay
#> intensity  symmetry 
#>  2.117293 41.071535 
#> 
#> $Woodcock
#>  strength     shape 
#>  1.490079 41.071535 
#> 
#> $Watterson_intensity
#> [1] 4.319083
#> 
#> $Lisle_intensity
#> [1] 3.67315
#> 
#> $Nadai
#>     goct     eoct 
#> 1.388465 1.202446 
#> 
#> $Lode
#> [1] -0.9524619
#> 
#> $kind
#> [1] "L"
#> 
#> $MAD_approx
#> [1] 17.97803
#> 
#> $MAD
#> [1] 17.97803
#>