List of vectors

Creates a list of Cartesian vectors from an spherical objects. This is a convenience function to link with the package geologyGeometry by J. R. Davis

Usage

vec_list(x)

Arguments

x

object of class "Vec3", "Line", "Ray", or "Plane", where the rows are the observations and the columns are the coordinates.

Value

list

Examples

vec_list(example_lines)
#> [[1]]
#>         x         y         z 
#> 0.5727204 0.7882819 0.2249511 
#> 
#> [[2]]
#>         x         y         z 
#> 0.4682901 0.8448178 0.2588190 
#> 
#> [[3]]
#>         x         y         z 
#> 0.2674497 0.9327081 0.2419219 
#> 
#> [[4]]
#>         x         y         z 
#> 0.1641876 0.9311540 0.3255682 
#> 
#> [[5]]
#>         x         y         z 
#> 0.4341533 0.8520738 0.2923717 
#> 
#> [[6]]
#>         x         y         z 
#> 0.2382466 0.9555548 0.1736482 
#> 
#> [[7]]
#>         x         y         z 
#> 0.2487928 0.9285075 0.2756374 
#> 
#> [[8]]
#>         x         y         z 
#> 0.3807167 0.8969124 0.2249511 
#> 
#> [[9]]
#>         x         y         z 
#> 0.4234340 0.8681684 0.2588190 
#> 
#> [[10]]
#>         x         y         z 
#> 0.5039832 0.8065417 0.3090170 
#> 
#> [[11]]
#>         x         y         z 
#> 0.5055754 0.8414187 0.1908090 
#> 
#> [[12]]
#>         x         y         z 
#> 0.6608925 0.7339955 0.1564345 
#> 
#> [[13]]
#>         x         y         z 
#> 0.1866394 0.9601763 0.2079117 
#> 
#> [[14]]
#>         x         y         z 
#> 0.3716074 0.8754521 0.3090170 
#> 
#> [[15]]
#>         x         y         z 
#> 0.3010969 0.9266810 0.2249511 
#> 
#> [[16]]
#>          x          y          z 
#> 0.06634235 0.94873979 0.30901699 
#> 
#> [[17]]
#>         x         y         z 
#> 0.2500000 0.9330127 0.2588190 
#> 
#> [[18]]
#>          x          y          z 
#> 0.03436929 0.98420783 0.17364818 
#> 
#> [[19]]
#>         x         y         z 
#> 0.1889521 0.9720741 0.1391731 
#> 
#> [[20]]
#>          x          y          z 
#> -0.1503745  0.9494270  0.2756374 
#> 
#> [[21]]
#>         x         y         z 
#> 0.2879299 0.9417763 0.1736482 
#> 
#> [[22]]
#>         x         y         z 
#> 0.3520151 0.8712678 0.3420201 
#> 
#> [[23]]
#>         x         y         z 
#> 0.3394708 0.9326883 0.1218693 
#> 
#> [[24]]
#>         x         y         z 
#> 0.1503745 0.9494270 0.2756374 
#> 
#> [[25]]
#>         x         y         z 
#> 0.3978485 0.8935823 0.2079117 
#> 
#> [[26]]
#>         x         y         z 
#> 0.2500000 0.9330127 0.2588190 
#> 
#> [[27]]
#>         x         y         z 
#> 0.4682901 0.8448178 0.2588190 
#> 
#> [[28]]
#>          x          y          z 
#> 0.48362865 0.87248918 0.06975647 
#> 
#> [[29]]
#>         x         y         z 
#> 0.3233864 0.8884968 0.3255682 
#> 
#> [[30]]
#>         x         y         z 
#> 0.3647789 0.8593653 0.3583679 
#> 
#> [[31]]
#>           x           y           z 
#> -0.05004912  0.95499417  0.29237170 
#> 
#> [[32]]
#>          x          y          z 
#> -0.1743127  0.8967611  0.4067366 
#> 
#> [[33]]
#>          x          y          z 
#> -0.1879102  0.9667141  0.1736482 
#> 
#> [[34]]
#>         x         y         z 
#> 0.2537255 0.8848461 0.3907311 
#> 
#> [[35]]
#>         x         y         z 
#> 0.2903810 0.8937008 0.3420201 
#> 
#> [[36]]
#>         x         y         z 
#> 0.1715103 0.9726831 0.1564345 
#> 
#> [[37]]
#>          x          y          z 
#> -0.4659722  0.8763662  0.1218693 
#> 
#> [[38]]
#>           x           y           z 
#> -0.20740523  0.97576488  0.06975647 
#> 
#> [[39]]
#>           x           y           z 
#> -0.06595604  0.94321534  0.32556815 
#> 
#> [[40]]
#>         x         y         z 
#> 0.4401396 0.5840850 0.6819984 
#> 
#> [[41]]
#>         x         y         z 
#> 0.4393867 0.7926747 0.4226183 
#> 
#> [[42]]
#>         x         y         z 
#> 0.3731498 0.8002223 0.4694716 
#> 
#> [[43]]
#>          x          y          z 
#> 0.58769573 0.80889378 0.01745241 
#> 
#> [[44]]
#>          x          y          z 
#> 0.62692564 0.77418869 0.08715574 
#> 
#> [[45]]
#>           x           y           z 
#> -0.30784109 -0.94743746  0.08715574 
#> 
#> [[46]]
#>          x          y          z 
#> -0.2548870 -0.9512512  0.1736482 
#> 
#> [[47]]
#>         x         y         z 
#> 0.5155271 0.7362490 0.4383711 
#> 
#> [[48]]
#>          x          y          z 
#> 0.06554964 0.93740358 0.34202014 
#> 
#> [[49]]
#>         x         y         z 
#> 0.1450435 0.9157687 0.3746066 
#> 
#> [[50]]
#>         x         y         z 
#> 0.4271357 0.8757580 0.2249511 
#> 
#> [[51]]
#>          x          y          z 
#> -0.4148536 -0.8896564  0.1908090 
#> 
#> [[52]]
#>          x          y          z 
#> -0.4341049 -0.8900470  0.1391731 
#> 
#> [[53]]
#>         x         y         z 
#> 0.4203031 0.9013429 0.1045285 
#> 
#> [[54]]
#>          x          y          z 
#> 0.45336832 0.88978543 0.05233596 
#> 
#> [[55]]
#>          x          y          z 
#> 0.39067162 0.92036466 0.01745241 
#> 
#> [[56]]
#>         x         y         z 
#> 0.8054738 0.4839775 0.3420201 
#> 
#> [[57]]
#>         x         y         z 
#> 0.1634922 0.5347588 0.8290376 
#> 
#> [[58]]
#>         x         y         z 
#> 0.5025704 0.5781411 0.6427876 
#> 
#> [[59]]
#>          x          y          z 
#> -0.3247917  0.8461112  0.4226183 
#> 
#> [[60]]
#>         x         y         z 
#> 0.3771196 0.8470246 0.3746066 
#> 
#> [[61]]
#>          x          y          z 
#>  0.3847950 -0.9065203  0.1736482 
#> 
#> [[62]]
#>         x         y         z 
#> 0.3059340 0.8884968 0.3420201 
#> 
#> [[63]]
#>         x         y         z 
#> 0.5504160 0.6797072 0.4848096 
#> 
#> [[64]]
#>         x         y         z 
#> 0.5720614 0.4156269 0.7071068 
#> 
#> [[65]]
#>          x          y          z 
#> -0.1165442 -0.9491766  0.2923717 
#> 
#> [[66]]
#>         x         y         z 
#> 0.3171156 0.8712678 0.3746066 
#> 
#> [[67]]
#>         x         y         z 
#> 0.3909804 0.8781563 0.2756374 
#> 
#> [[68]]
#>         x         y         z 
#> 0.3755949 0.8848461 0.2756374 
#> 
#> [[69]]
#>         x         y         z 
#> 0.5072135 0.8441450 0.1736482 
#> 
#> [[70]]
#>         x         y         z 
#> 0.3099755 0.8516507 0.4226183 
#> 
#> [[71]]
#>         x         y         z 
#> 0.3878187 0.9136435 0.1218693 
#> 
#> [[72]]
#>         x         y         z 
#> 0.4698463 0.8137977 0.3420201 
#> 
#> [[73]]
#>         x         y         z 
#> 0.5785017 0.7676977 0.2756374 
#> 
#> [[74]]
#>         x         y         z 
#> 0.3184537 0.9248567 0.2079117 
#> 
#> [[75]]
#>         x         y         z 
#> 0.2573296 0.8974151 0.3583679 
#> 
#> [[76]]
#>          x          y          z 
#> -0.1495990 -0.9445311  0.2923717 
#> 
#> [[77]]
#>         x         y         z 
#> 0.5681638 0.6535977 0.5000000 
#> 
#> [[78]]
#>         x         y         z 
#> 0.4723840 0.8522033 0.2249511 
#> 
#> [[79]]
#>         x         y         z 
#> 0.4232640 0.6517694 0.6293204 
#> 
#> [[80]]
#>          x          y          z 
#> -0.7875307 -0.6152864  0.0348995 
#> 
#> [[81]]
#>          x          y          z 
#> 0.48473578 0.87448650 0.01745241 
#> 
#> [[82]]
#>         x         y         z 
#> 0.6049416 0.7470406 0.2756374 
#> 
#> [[83]]
#>         x         y         z 
#> 0.6815829 0.7309082 0.0348995 
#> 
#> [[84]]
#>         x         y         z 
#> 0.5677570 0.7814504 0.2588190 
#>