Simple fault analysis

The PT-techniques is a graphical solution of the Wallace-Bott hypothesis, i.e. fault slip occurs parallel to the maximum shear stress. It calculates PT-axes, kinematic planes (also movement planes), and the dihedra separation plane.

Usage

Fault_PT(x, ptangle = 90)

Arguments

x

object of class "Fault"

ptangle

numeric. angle between P and T axes in degrees (90° by default).

Value

list. p and t are the P and T axes as "Line" objects, m and d are the M-planes and the dihedra separation planes as "Plane" objects

See also

Examples

f <- Fault(c(120, 120, 100), c(60, 60, 50), c(110, 25, 30), c(58, 9, 23), c(1, -1, 1))
Fault_PT(f)
#> $p
#> Line object (n = 3):
#>       azimuth   plunge
#> [1,] 314.9694 75.19695
#> [2,] 248.4032 15.30354
#> [3,] 342.4517 46.65113
#> 
#> $t
#> Line object (n = 3):
#>       azimuth   plunge
#> [1,] 116.2068 14.04837
#> [2,] 345.9417 25.61490
#> [3,] 241.3308 10.31892
#> 
#> $m
#> Plane object (n = 3):
#>      dip_direction      dip
#> [1,]      27.35344 85.42739
#> [2,]     310.64119 30.43222
#> [3,]     322.06010 48.49732
#> 
#> $d
#> Plane object (n = 3):
#>      dip_direction      dip
#> [1,]      289.7677 31.20915
#> [2,]      208.8622 83.16168
#> [3,]      210.2233 67.19865
#>