Mohr circle parameters

calculates parameters of the Mohr circle

Usage

Mohr_calc(
  sigma_x = NA,
  sigma_z = NA,
  tau_xz = NA,
  sigma1 = NA,
  sigma3 = NA,
  theta = seq(from = 0, to = 180, by = 1)
)

Arguments

sigma_x

numeric. Magnitude of normal stress acting in the horizontal direction

sigma_z

numeric. Magnitude of normal stress acting in the vertical direction

tau_xz

numeric. Magnitude of shear stress acting on the same plane as "sigma_x"

sigma1

numeric. Magnitude of major principal stress

sigma3

numeric. Magnitude of minor principal stress

theta

numeric. Angles (degrees); defaults to 0-180 in increments of 1

Value

A five-element list containing

"mean"

center of Mohr circle, i.e. the mean stress

"deviatoric"

radius of Mohr circle, i.e. the deviatoric stress

"sigma"

vector of normal stresses for Mohr circle

"tau"

vector of shear stresses for Mohr circle

"theta"

vector of angles (degrees)

Details

If theta is entered in conjunction with "sigma_x", "sigma_z", and "tau_xz", it is interpreted as the angle of inclination above the horizontal. If "theta" is entered in conjunction with the principal stresses, then it is interpreted as the angle of inclination above the major principal plane.

Note

One of the following two sets of data must be entered

"sigma_x", "sigma_z", "tau_xz"

"sigma1", "sigma3"

See also

Author

Kyle Elmy and Jim Kaklamanos

Examples

Mohr_calc(sigma_x = 80, sigma_z = 120, tau_xz = 20)
#> $mean
#> [1] 100
#> 
#> $deviatoric
#> [1] 28.28427
#> 
#> $normal
#>   [1] 120.00000 120.68581 121.34641 121.98101 122.58882 123.16912 123.72119
#>   [8] 124.24435 124.73798 125.20147 125.63426 126.03581 126.40564 126.74330
#>  [15] 127.04838 127.32051 127.55935 127.76461 127.93604 128.07344 128.17664
#>  [22] 128.24551 128.27996 128.27996 128.24551 128.17664 128.07344 127.93604
#>  [29] 127.76461 127.55935 127.32051 127.04838 126.74330 126.40564 126.03581
#>  [36] 125.63426 125.20147 124.73798 124.24435 123.72119 123.16912 122.58882
#>  [43] 121.98101 121.34641 120.68581 120.00000 119.28983 118.55615 117.79987
#>  [50] 117.02190 116.22319 115.40472 114.56748 113.71249 112.84079 111.95345
#>  [57] 111.05155 110.13618 109.20846 108.26952 107.32051 106.36258 105.39689
#>  [64] 104.42463 103.44699 102.46514 101.48028 100.49363  99.50637  98.51972
#>  [71]  97.53486  96.55301  95.57537  94.60311  93.63742  92.67949  91.73048
#>  [78]  90.79154  89.86382  88.94845  88.04655  87.15921  86.28751  85.43252
#>  [85]  84.59528  83.77681  82.97810  82.20013  81.44385  80.71017  80.00000
#>  [92]  79.31419  78.65359  78.01899  77.41118  76.83088  76.27881  75.75565
#>  [99]  75.26202  74.79853  74.36574  73.96419  73.59436  73.25670  72.95162
#> [106]  72.67949  72.44065  72.23539  72.06396  71.92656  71.82336  71.75449
#> [113]  71.72004  71.72004  71.75449  71.82336  71.92656  72.06396  72.23539
#> [120]  72.44065  72.67949  72.95162  73.25670  73.59436  73.96419  74.36574
#> [127]  74.79853  75.26202  75.75565  76.27881  76.83088  77.41118  78.01899
#> [134]  78.65359  79.31419  80.00000  80.71017  81.44385  82.20013  82.97810
#> [141]  83.77681  84.59528  85.43252  86.28751  87.15921  88.04655  88.94845
#> [148]  89.86382  90.79154  91.73048  92.67949  93.63742  94.60311  95.57537
#> [155]  96.55301  97.53486  98.51972  99.50637 100.49363 101.48028 102.46514
#> [162] 103.44699 104.42463 105.39689 106.36258 107.32051 108.26952 109.20846
#> [169] 110.13618 111.05155 111.95345 112.84079 113.71249 114.56748 115.40472
#> [176] 116.22319 117.02190 117.79987 118.55615 119.28983 120.00000
#> 
#> $shear
#>   [1] -20.0000000 -19.2898266 -18.5561515 -17.7998686 -17.0218994 -16.2231915
#>   [7] -15.4047182 -14.5674766 -13.7124868 -12.8407904 -11.9534495 -11.0515452
#>  [13] -10.1361763  -9.2084580  -8.2695206  -7.3205081  -6.3625766  -5.3968934
#>  [19]  -4.4246348  -3.4469856  -2.4651367  -1.4802844  -0.4936286   0.4936286
#>  [25]   1.4802844   2.4651367   3.4469856   4.4246348   5.3968934   6.3625766
#>  [31]   7.3205081   8.2695206   9.2084580  10.1361763  11.0515452  11.9534495
#>  [37]  12.8407904  13.7124868  14.5674766  15.4047182  16.2231915  17.0218994
#>  [43]  17.7998686  18.5561515  19.2898266  20.0000000  20.6858065  21.3464105
#>  [49]  21.9810072  22.5888234  23.1691186  23.7211858  24.2443524  24.7379810
#>  [55]  25.2014702  25.6342553  26.0358090  26.4056420  26.7433039  27.0483831
#>  [61]  27.3205081  27.5593472  27.7646095  27.9360449  28.0734446  28.1766411
#>  [67]  28.2455086  28.2799634  28.2799634  28.2455086  28.1766411  28.0734446
#>  [73]  27.9360449  27.7646095  27.5593472  27.3205081  27.0483831  26.7433039
#>  [79]  26.4056420  26.0358090  25.6342553  25.2014702  24.7379810  24.2443524
#>  [85]  23.7211858  23.1691186  22.5888234  21.9810072  21.3464105  20.6858065
#>  [91]  20.0000000  19.2898266  18.5561515  17.7998686  17.0218994  16.2231915
#>  [97]  15.4047182  14.5674766  13.7124868  12.8407904  11.9534495  11.0515452
#> [103]  10.1361763   9.2084580   8.2695206   7.3205081   6.3625766   5.3968934
#> [109]   4.4246348   3.4469856   2.4651367   1.4802844   0.4936286  -0.4936286
#> [115]  -1.4802844  -2.4651367  -3.4469856  -4.4246348  -5.3968934  -6.3625766
#> [121]  -7.3205081  -8.2695206  -9.2084580 -10.1361763 -11.0515452 -11.9534495
#> [127] -12.8407904 -13.7124868 -14.5674766 -15.4047182 -16.2231915 -17.0218994
#> [133] -17.7998686 -18.5561515 -19.2898266 -20.0000000 -20.6858065 -21.3464105
#> [139] -21.9810072 -22.5888234 -23.1691186 -23.7211858 -24.2443524 -24.7379810
#> [145] -25.2014702 -25.6342553 -26.0358090 -26.4056420 -26.7433039 -27.0483831
#> [151] -27.3205081 -27.5593472 -27.7646095 -27.9360449 -28.0734446 -28.1766411
#> [157] -28.2455086 -28.2799634 -28.2799634 -28.2455086 -28.1766411 -28.0734446
#> [163] -27.9360449 -27.7646095 -27.5593472 -27.3205081 -27.0483831 -26.7433039
#> [169] -26.4056420 -26.0358090 -25.6342553 -25.2014702 -24.7379810 -24.2443524
#> [175] -23.7211858 -23.1691186 -22.5888234 -21.9810072 -21.3464105 -20.6858065
#> [181] -20.0000000
#> 
#> $theta
#>   [1]   0   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17
#>  [19]  18  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35
#>  [37]  36  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53
#>  [55]  54  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71
#>  [73]  72  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89
#>  [91]  90  91  92  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107
#> [109] 108 109 110 111 112 113 114 115 116 117 118 119 120 121 122 123 124 125
#> [127] 126 127 128 129 130 131 132 133 134 135 136 137 138 139 140 141 142 143
#> [145] 144 145 146 147 148 149 150 151 152 153 154 155 156 157 158 159 160 161
#> [163] 162 163 164 165 166 167 168 169 170 171 172 173 174 175 176 177 178 179
#> [181] 180
#>