Fault-Slip Data

List of different fault-slip data from Angelier (1990)

Usage

data('angelier1990')

Format

An object of class data.frame

References

Angelier, J. (1990). Inversion of field data in fault tectonics to obtain the regional stress-III. A new rapid direct inversion method by analytical means. Geophysical Journal International, 103(2), 363–376. https://doi.org/10.1111/j.1365-246X.1990.tb01777.x

Examples

data("angelier1990")
head(angelier1990)
#> $AVB
#> Fault object (n = 33):
#>       dip_direction dip    azimuth   plunge sense
#>  [1,]           329  71  10.372108 65.35251     1
#>  [2,]           315  65 357.248859 57.79119     1
#>  [3,]           312  70   1.751099 60.60630     1
#>  [4,]           317  63 330.035014 62.39035     1
#>  [5,]           134  56 146.384168 55.37193     1
#>  [6,]           162  65 141.455533 63.52766     1
#>  [7,]           153  70 128.151847 68.14416     1
#>  [8,]           316  61 330.212091 60.23864     1
#>  [9,]           340  73 356.659170 72.30080     1
#> [10,]           342  72 338.767039 71.97316     1
#> [11,]           323  86  40.883188 71.57529     1
#> [12,]           330  74  21.322413 65.35251     1
#> [13,]           317  77  24.911582 58.45212     1
#> [14,]           325  65  18.796010 51.71010     1
#> [15,]           165  52 124.184821 44.08878     1
#> [16,]           309  85  24.795705 70.37664     1
#> [17,]           312  71   1.697588 61.97195     1
#> [18,]           329  63   6.178380 57.40115     1
#> [19,]           319  58 354.918900 52.34664     1
#> [20,]           317  57 338.319236 55.11920     1
#> [21,]           174  56 136.798420 49.74119     1
#> [22,]           319  40 346.808006 36.58269     1
#> [23,]           159  51 121.223788 44.30632     1
#> [24,]           167  32 140.410590 29.19564     1
#> [25,]           322  54  11.988412 41.50672     1
#> [26,]           149  67 110.958314 61.67654     1
#> [27,]           317  65   9.677250 52.43617     1
#> [28,]           151  86  70.332342 66.67339     1
#> [29,]           317  74  29.406985 46.50824     1
#> [30,]           301  69  11.283948 41.31063     1
#> [31,]           149  65 160.695922 64.53645     1
#> [32,]           146  56 126.832105 54.46945     1
#> [33,]           340  80 334.259926 79.95063     1
#> 
#> $TYM
#> Fault object (n = 38):
#>       dip_direction dip    azimuth   plunge sense
#>  [1,]           137  61 117.013494 59.46650     1
#>  [2,]           128  59 146.899008 57.58045     1
#>  [3,]             2  80 287.530355 56.63295     1
#>  [4,]           324  68 294.428870 65.08349     1
#>  [5,]           317  63 291.911192 60.63761     1
#>  [6,]            22  88 295.324148 58.94196     1
#>  [7,]           346  78 280.030566 62.43703     1
#>  [8,]           138  60 157.425400 58.52505     1
#>  [9,]           349  61 357.207488 60.74920     1
#> [10,]           159  56 155.426604 55.94829     1
#> [11,]           141  70 189.299188 61.31581     1
#> [12,]           308  50 314.208610 49.83395     1
#> [13,]           150  51 167.164526 49.71766     1
#> [14,]           171  62 197.186198 59.35227     1
#> [15,]           328  62 315.380833 61.41529     1
#> [16,]           306  61 277.071031 57.65242     1
#> [17,]           126  60 152.503414 57.17189     1
#> [18,]           129  63 139.907773 62.57518     1
#> [19,]           160  72 101.235974 57.92973     1
#> [20,]           141  53 108.468482 48.20984     1
#> [21,]           281  47 296.908389 45.88274     1
#> [22,]           329  45 298.023689 40.60913     1
#> [23,]           204  74 262.238697 61.42075     1
#> [24,]           297  42 298.345522 41.99214     1
#> [25,]           306  56 311.354144 55.88376     1
#> [26,]           129  77 203.992675 48.28042     1
#> [27,]           149  61 196.641620 50.55577     1
#> [28,]           340  58   8.418328 54.60676     1
#> [29,]           153  67 175.065372 65.39117     1
#> [30,]           120  58 169.700452 45.98728     1
#> [31,]           122  69 108.280689 68.43916     1
#> [32,]           133  63 146.035014 62.39035     1
#> [33,]           115  68 107.035962 67.80646     1
#> [34,]           341  48 348.449152 47.75842     1
#> [35,]           340  69 334.434447 68.90926     1
#> [36,]           287  68 312.206323 65.93702     1
#> [37,]             6  70 322.468751 63.34203     1
#> [38,]           359  71   2.068927 70.97467     1
#>