Structure classes

Line, Plane, and Fault create a "line", "plane", "pair", and "fault" class object, respectively, from the given set of values.

as.line, as.plane, as.pair, and as.fault attempt to turn its argument into a "line", "plane", and "fault" class object, respectively.

is.line, is.plane, is.pair, and is.fault test if its argument is a "line", "plane", and "fault" class object, respectively.

Usage

Line(azimuth, plunge)

Plane(dip_direction, dip)

Fault(dip_direction, dip, azimuth, plunge, sense = NULL)

Pair(dip_direction, dip, azimuth, plunge)

as.line(l)

as.plane(p)

as.fault(f)

as.pair(f)

is.line(l)

is.plane(p)

is.fault(f)

is.pair(f)

is.spherical(l)

Arguments

azimuth, plunge

numeric vectors. Azimuth and plunge of a line (in degrees)

dip_direction, dip

numeric vectors. Dip direction and dip of a plane (in degrees)

sense

(optional) integer. Sense of the line on a fault plane. Either 1or -1 for normal or thrust offset, respectively.

l, p, f

numeric vector or array containing the spherical coordinates (1st element/column is azimuth, 2nd element/column is inclination, both in degrees), or object of class "line", "plane", "pair", or "fault"

Details

is.line, is.plane, is.pair, and is.fault return TRUE if l, p, and f are an object of class "line", "plane", "pair" or "fault", respectively, and FALSE otherwise.

is.spherical returns TRUE if the argument's class is one of "line", "plane", "pair", or "fault" and FALSE otherwise

as.line, as.plane, as.pair, and as.fault are is generic functions. If the argument is a "line" or "plane" class, it will be converted.

Examples

x <- Line(120, 50) # create line
is.line(x) # test if line
#> [1] TRUE
as.plane(x) # convert to plane
#>      dip_direction dip
#> [1,]           300  40
#> attr(,"class")
#> [1] "plane"