Define a Map (Object) Structure using Named Arguments

Creates an intermediate definition for a JSON object (map) with named fields and their corresponding structure definitions. This definition is used within s_vector() or finalized by build_structure().

Usage

s_map(..., .ignore_extra_fields = FALSE)

Arguments

...

Named arguments defining the map's structure. The names correspond to the required JSON object keys (strings). The values must be structure definitions created by s_* function calls (e.g., id = s_integer(), tags = s_vector(s_string()), metadata = s_map(...)). An empty map (representing an empty JSON object {}) can be created by calling s_map() with no arguments.

.ignore_extra_fields

Logical (default FALSE). If TRUE, allows the JSON object to contain fields not specified in .... These extra fields will be ignored during parsing. If FALSE (default), extra fields in the JSON will cause a validation error. Note: This argument is currently processed by build_structure.

Value

An intermediate list representing the map structure definition.

Examples

# Define a simple map with a string 'name' and an integer 'age'
map_def1 <- s_map(name = s_string(), age = s_integer())
build_structure(map_def1)
#> <pointer: 0x556db4c31680>
#> attr(,"class")
#> [1] "Structure"
# Expected: list(type = "map", value = list(name = list(type = "string"),
#                                          age = list(type = "integer")),
#                ignore_extra_fields = FALSE)

# Define a nested map
map_def2 <- s_map(
  user = s_string(),
  details = s_map(
    email = s_string(),
    active = s_logical()
  ),
  permissions = s_vector(s_string())
)
str(build_structure(map_def2)) # Use str() for compact view
#> Class 'Structure' <externalptr> 

# Define an empty map
empty_map_def <- s_map()
build_structure(empty_map_def)
#> <pointer: 0x556db4d37fd0>
#> attr(,"class")
#> [1] "Structure"
# Expected: list(type = "map", value = list(), ignore_extra_fields = FALSE)

# Define a map ignoring extra fields
map_def_ignore <- s_map(required_field = s_integer(), .ignore_extra_fields = TRUE)
build_structure(map_def_ignore)
#> <pointer: 0x556db519c320>
#> attr(,"class")
#> [1] "Structure"
# Expected: list(type = "map", value = list(required_field = list(type = "integer")),
#                ignore_extra_fields = TRUE)

# Parsing example (see ?parse_json)
json_data <- '{"name": "Example", "age": 42}'
parsed <- parse_json(json_data, build_structure(map_def1))
print(parsed) # list(name = "Example", age = 42L)
#> $age
#> [1] 42
#> 
#> $name
#> [1] "Example"
#> 

json_extra <- '{"name": "Example", "age": 42, "extra": true}'
# This will error by default:
try(parse_json(json_extra, build_structure(map_def1)))
#> Error in parse_json(json_extra, build_structure(map_def1)) : 
#>   Serde("unknown field `extra`, expected `name` or `age`") at character 0
# This will succeed if using ignore_extra_fields = TRUE definition:
parsed_ignore <- parse_json(json_extra, build_structure(map_def_ignore))
#> Error in parse_json(json_extra, build_structure(map_def_ignore)): Serde("missing field `required_field`") at character 0
# print(parsed_ignore) # list(required_field = ...) - requires matching json